refactor : modularise editor store
