+++ lib/editor/store.tsx
@@ -2,768 +2,17 @@
 
 import { create } from "zustand"
 
-import { DEFAULT_CLIP_DURATION_MS } from "./animation-motion"
-import { NEW_CLIP_EASING } from "./clip-easing"
-import { NEUTRAL_MEDIA_ADJUSTMENTS } from "./css-utils"
-import {
-  clipAffectsMain,
-  clipAffectsSlot,
-  clipBaseline,
-  clipPose,
-  DEFAULT_BASELINE,
-  poseAtCut,
-  REST_LIGHTING,
-} from "./animation-playback"
-import { MAX_DURATION_MS, resolveRippleDrop } from "./animation-timeline"
-import { isVideoSrc } from "./media-type"
-import { LAYOUT_PRESETS, PRESENT_PRESETS } from "./present-presets"
-import { screenshotPositionAnchor } from "./presets"
-import { mergeCanvasStyle } from "./preset-fields"
-import type { CustomPresetCanvasStyle } from "./preset-fields"
-import type { PresetSort } from "@/lib/schemas/preset"
-import {
-  resolveActivePresetGeometry,
-  resolveMainOffsetPx,
-  resolveSlotPositionPct,
-} from "./preset-geometry"
-import { computeRowLayout } from "./screenshot-layout"
-import { computeNextLayerZ, moveLayerInStack } from "./store/layer-stack"
-import {
-  applyScreenshotStyle,
-  applySharedFrameToCanvas,
-  aspectRatioFromState,
-  CANVAS_BASE_W,
-  clampPct,
-  cloneBorder,
-  cloneLighting,
-  cloneShadow,
-  createScreenshotSlot,
-  duplicateLayerItem,
-  layoutSlotsInRow,
-  makeId,
-  placeNewSlotInRow,
-  placementAfterCanvas,
-  removeSlotFromRow,
-  resolveActiveLayoutGeometry,
-  resolveMainScreenshotStyle,
-  resolveSlotScreenshotStyle,
-  screenshotStyleEffects,
-  screenshotStyleGroup,
-  type ScreenshotStylePatch,
-  type ScreenshotStyleTarget,
-  scaleAnnotationStrokesForAspectChange,
-  scaleScreenshotOffsetForAspectChange,
-  stateCanvasAspect,
-  type PresetTab,
-} from "./store/canvas-helpers"
-import {
-  CLEAR_SELECTION,
-  createCanvas,
-  DEFAULT_STATE,
-  GROUP_MERGE_MS,
-  HISTORY_LIMIT,
-  MAX_CANVASES,
-  MAX_SCREENSHOT_SLOTS,
-} from "./store/defaults"
-import {
-  normalizeEditorState,
-  type CurrentDraftInfo,
-} from "./store/draft-persistence"
-import { remapAnimationForApply } from "./custom-preset-snapshot"
-import { FONT_FAMILIES } from "./fonts"
-import type {
-  Annotation,
-  AnnotationPoint,
-  AnnotationShape,
-  AnnotationStroke,
-  AnimationClip,
-  AnimationClipTarget,
-  AnimationEffect,
-  ClipBaseline,
-  ClipSlotPose,
-  CanvasAnimation,
-  AspectState,
-  AssetElement,
-  AssetFilter,
-  Background,
-  BackdropEffects,
-  BackdropLighting,
-  BackdropPattern,
-  Border,
-  CanvasState,
-  CropRegion,
-  DeviceFrame,
-  EditorState,
-  EditorTool,
-  EnhancePreset,
-  MediaAdjustments,
-  Overlay,
-  Portrait,
-  ScreenshotLayer,
-  ScreenshotPosition,
-  ScreenshotSlot,
-  Shadow,
-  TextElement,
-  Tilt,
-  TweetCard,
-  VideoTimelineClip,
-} from "./state-types"
-
-const TWEET_POST_ASPECT: AspectState = { id: "x-post", w: 1080, h: 1080 }
-
-/** Smallest an animation clip may be trimmed/fitted to (keep in sync with the
- * timeline UI's MIN_CLIP_MS). */
-const MIN_ANIMATION_CLIP_MS = 200
-
-/** Effects an extra screenshot slot can actually animate (its keyframe pose only
- * carries transform + shadow). Used to decide when editing an effect may auto-
- * bind an unbound keyframe to a slot. */
-const SLOT_ANIMATABLE_EFFECTS: AnimationEffect[] = [
-  "tilt",
-  "zoom",
-  "shadow",
-  "position",
-  "border",
-  "borderRadius",
-  "padding",
-  "lighting",
-  "mediaEffects",
-  "mediaFilter",
-]
-
-/** Canvas.animation is optional (older drafts) — always read through this. */
-const getCanvasAnimation = (canvas: CanvasState): CanvasAnimation =>
-  canvas.animation ?? { durationMs: 5000, clips: [] }
-
-/** Snapshot the canvas's animatable state as a clip's target keyframe (pose). */
-export const captureClipPose = (canvas: CanvasState): ClipBaseline => {
-  // Main and slot style are projected through the same resolver pair so the two
-  // branches of the pose stay structurally identical.
-  const main = resolveMainScreenshotStyle(canvas)
-  return {
-    tilt: main.tilt,
-    scale: main.scale,
-    screenshotPosition: canvas.screenshotPosition,
-    screenshotOffset: canvas.screenshotOffset,
-    padding: main.padding,
-    canvasBorderRadius: canvas.canvasBorderRadius,
-    shadow: main.shadow,
-    backdropEffects: canvas.backdrop.effects,
-    lighting: main.lighting,
-    background: canvas.background,
-    filter: canvas.backdrop.filter,
-    mediaAdjustments: main.adjustments,
-    mediaFilter: main.filter,
-    portrait: canvas.portrait,
-    pattern: canvas.backdrop.pattern,
-    overlay: canvas.overlay,
-    border: main.border,
-    borderRadius: main.borderRadius,
-    // null (not undefined) so a clip with no crop stays distinguishable from a
-    // pose captured before crop animated — see poseCrop.
-    crop: canvas.lastCropRegion,
-    slots: Object.fromEntries(
-      canvas.screenshotSlots.map((s) => {
-        // resolveSlotScreenshotStyle is the single owner of the slot→canvas
-        // inheritance rule (a slot with no shadow/border/… borrows the canvas's).
-        const style = resolveSlotScreenshotStyle(s, canvas)
-        return [
-          s.id,
-          {
-            tilt: style.tilt,
-            scale: style.scale,
-            rotation: s.rotation,
-            shadow: style.shadow,
-            xPct: s.xPct,
-            yPct: s.yPct,
-            border: style.border,
-            borderRadius: style.borderRadius,
-            padding: style.padding,
-            lighting: style.lighting,
-            filter: style.filter,
-            adjustments: style.adjustments,
-          },
-        ]
-      })
-    ),
-  }
-}
-
-/**
- * Resolve a pose's crop against the live one. Poses captured before crop
- * animated omit the key entirely (undefined) and inherit the live crop; an
- * explicit null means this clip has no crop and must NOT inherit one, or
- * clearing a crop would come back the moment another clip applied one.
- */
-const poseCrop = (pose: ClipBaseline, live: CropRegion | null) =>
-  pose.crop !== undefined ? pose.crop : live
-
-/**
- * Load a clip's pose onto the canvas's live/committed style so the inspector and
- * canvas show that clip's keyframe. Slots not present in the pose keep their
- * current transform.
- */
-const applyPoseToCanvas = (
-  canvas: CanvasState,
-  pose: ClipBaseline
-): Partial<CanvasState> => ({
-  tilt: pose.tilt,
-  scale: pose.scale,
-  screenshotPosition: pose.screenshotPosition,
-  screenshotOffset: pose.screenshotOffset,
-  padding: pose.padding,
-  // Fall back to the live value for poses captured before this field existed.
-  canvasBorderRadius: pose.canvasBorderRadius ?? canvas.canvasBorderRadius,
-  shadow: pose.shadow,
-  background: pose.background,
-  // Fall back to the live value for poses captured before portrait animated.
-  portrait: pose.portrait ?? canvas.portrait,
-  // Fall back to the live value for poses captured before overlay animated.
-  overlay: pose.overlay ?? canvas.overlay,
-  // Fall back to the live value for poses captured before border animated.
-  border: pose.border ?? canvas.border,
-  // Fall back to the live value for poses captured before radius animated.
-  borderRadius: pose.borderRadius ?? canvas.borderRadius,
-  // Fall back to the live values for poses captured before the media grade
-  // animated.
-  mediaAdjustments: pose.mediaAdjustments ?? canvas.mediaAdjustments,
-  mediaFilter: pose.mediaFilter ?? canvas.mediaFilter,
-  lastCropRegion: poseCrop(pose, canvas.lastCropRegion),
-  backdrop: {
-    ...canvas.backdrop,
-    effects: pose.backdropEffects,
-    // Fall back to the live value for poses captured before lighting animated.
-    lighting: pose.lighting ?? canvas.backdrop.lighting,
-    // Fall back to the live value for poses captured before filter animated.
-    filter: pose.filter ?? canvas.backdrop.filter,
-    // Fall back to the live value for poses captured before pattern animated.
-    pattern: pose.pattern ?? canvas.backdrop.pattern,
-  },
-  screenshotSlots: canvas.screenshotSlots.map((s) => {
-    const sp = pose.slots[s.id]
-    if (!sp) return s
-    return {
-      ...s,
-      tilt: sp.tilt,
-      scale: sp.scale,
-      rotation: sp.rotation,
-      // Only overwrite the slot's shadow when the pose captured one (older poses
-      // are transform-only, so leave the committed shadow untouched there).
-      ...(sp.shadow ? { shadow: sp.shadow } : {}),
-      // Likewise only restore position for poses that captured it (older poses
-      // omit it, so leave the committed position untouched there).
-      ...(sp.xPct != null && sp.yPct != null
-        ? { xPct: sp.xPct, yPct: sp.yPct }
-        : {}),
-      // Border / radius / padding / lighting: only restore when the pose captured
-      // them (older poses omit them → leave the committed slot values untouched).
-      ...(sp.border ? { border: sp.border } : {}),
-      ...(sp.borderRadius != null ? { borderRadius: sp.borderRadius } : {}),
-      ...(sp.padding != null ? { padding: sp.padding } : {}),
-      ...(sp.lighting ? { lighting: sp.lighting } : {}),
-      ...(sp.filter ? { filter: sp.filter } : {}),
-      ...(sp.adjustments ? { adjustments: sp.adjustments } : {}),
-    }
-  }),
-})
-
-/**
- * The pose the static/committed canvas holds while it carries an animation:
- * where the animation STARTS.
- *
- * Every effect resolves to the first keyframe that owns it — the value that
- * keyframe eases FROM — and an effect no keyframe owns keeps the canvas's own
- * committed value. So an animation stays a *description of motion* rather than
- * a style edit: leaving Animate mode never repaints the document with the last
- * keyframe's look, and the still/PNG export shows the composition the user
- * built, not the end of its animation. (Position always worked this way; the
- * rest used to rest at the final frame.)
- */
-const buildRestingPose = (
-  canvas: CanvasState,
-  clips: readonly AnimationClip[]
-): ClipBaseline | null => {
-  if (clips.length === 0) return null
-  const sorted = [...clips].sort((a, b) => a.startMs - b.startMs)
-  const committed = captureClipPose(canvas)
-  const rest: ClipBaseline = { ...committed, slots: { ...committed.slots } }
-  const restRec = rest as Record<string, unknown>
-
-  for (const effect of ANIMATION_EFFECTS) {
-    const firstMain = sorted.find(
-      (c) => clipAffectsMain(c) && clipOwnsEffect(c, effect)
-    )
-    if (firstMain) {
-      const base = clipBaseline(firstMain) as unknown as Record<string, unknown>
-      for (const field of EFFECT_MAIN_POSE_FIELDS[effect]) {
-        if (base[field] !== undefined) restRec[field] = base[field]
-      }
-    }
-
-    const slotFields = EFFECT_SLOT_POSE_FIELDS[effect]
-    if (!slotFields) continue
-    for (const slotId of Object.keys(rest.slots)) {
-      const firstSlot = sorted.find(
-        (c) => clipAffectsSlot(c, slotId) && clipOwnsEffect(c, effect)
-      )
-      if (!firstSlot) continue
-      const from = clipBaseline(firstSlot).slots[slotId] as
-        | Record<string, unknown>
-        | undefined
-      if (!from) continue
-      const next = { ...rest.slots[slotId] } as Record<string, unknown>
-      for (const field of slotFields) {
-        if (from[field] !== undefined) next[field] = from[field]
-      }
-      rest.slots[slotId] = next as unknown as ClipSlotPose
-    }
-  }
-
-  return rest
-}
-
-const clipOwnsEffect = (clip: AnimationClip, effect: AnimationEffect) =>
-  (clip.effects ?? []).includes(effect)
-
-/**
- * Copy `from`'s value for every effect `owns` selects onto a copy of `base`,
- * field by field (main pose and per-slot alike). `clips` is scanned for the
- * ownership test, so a caller can ask "whatever ANY of these clips animates" or
- * narrow it to one clip's own effects.
- */
-const overlayEffectFields = (
-  base: ClipBaseline,
-  from: ClipBaseline,
-  clips: readonly AnimationClip[],
-  owns: (clip: AnimationClip, effect: AnimationEffect) => boolean
-): ClipBaseline => {
-  const next: ClipBaseline = { ...base, slots: { ...base.slots } }
-  const nextRec = next as unknown as Record<string, unknown>
-  const fromRec = from as unknown as Record<string, unknown>
-
-  for (const effect of ANIMATION_EFFECTS) {
-    if (clips.some((c) => clipAffectsMain(c) && owns(c, effect))) {
-      for (const field of EFFECT_MAIN_POSE_FIELDS[effect]) {
-        if (fromRec[field] !== undefined) nextRec[field] = fromRec[field]
-      }
-    }
-    const slotFields = EFFECT_SLOT_POSE_FIELDS[effect]
-    if (!slotFields) continue
-    for (const slotId of Object.keys(next.slots)) {
-      if (!clips.some((c) => clipAffectsSlot(c, slotId) && owns(c, effect))) {
-        continue
-      }
-      const src = from.slots[slotId] as Record<string, unknown> | undefined
-      if (!src) continue
-      const merged = { ...next.slots[slotId] } as Record<string, unknown>
-      for (const field of slotFields) {
-        if (src[field] !== undefined) merged[field] = src[field]
-      }
-      next.slots[slotId] = merged as unknown as ClipSlotPose
-    }
-  }
-  return next
-}
-
-/** Main-pose fields each animation effect owns. Used to copy just the edited
- * property into every clip of a multi-selection (leaving the rest intact). */
-const EFFECT_MAIN_POSE_FIELDS: Record<
-  AnimationEffect,
-  readonly (keyof ClipBaseline)[]
-> = {
-  position: ["screenshotPosition", "screenshotOffset"],
-  zoom: ["scale"],
-  tilt: ["tilt"],
-  padding: ["padding"],
-  shadow: ["shadow"],
-  background: ["background"],
-  backdrop: ["backdropEffects"],
-  canvasRadius: ["canvasBorderRadius"],
-  lighting: ["lighting"],
-  filter: ["filter"],
-  mediaEffects: ["mediaAdjustments"],
-  mediaFilter: ["mediaFilter"],
-  portrait: ["portrait"],
-  pattern: ["pattern"],
-  overlay: ["overlay"],
-  border: ["border"],
-  borderRadius: ["borderRadius"],
-  crop: ["crop"],
-}
-
-const ANIMATION_EFFECTS = Object.keys(
-  EFFECT_MAIN_POSE_FIELDS
-) as AnimationEffect[]
-
-/** Per-slot pose fields an effect owns (only the slot-animatable ones). */
-const EFFECT_SLOT_POSE_FIELDS: Partial<
-  Record<AnimationEffect, readonly (keyof ClipSlotPose)[]>
-> = {
-  position: ["xPct", "yPct"],
-  zoom: ["scale"],
-  tilt: ["tilt", "rotation"],
-  padding: ["padding"],
-  shadow: ["shadow"],
-  border: ["border"],
-  borderRadius: ["borderRadius"],
-  lighting: ["lighting"],
-  mediaEffects: ["adjustments"],
-  mediaFilter: ["filter"],
-}
-
-const poseValueEq = (a: unknown, b: unknown): boolean =>
-  a === b || JSON.stringify(a) === JSON.stringify(b)
-
-/**
- * Copy just the fields owned by `effects` from `edited` onto `basePose`, so
- * applying one inspector edit across a multi-selection updates only that
- * property on every selected clip and leaves each clip's other keyframed values
- * intact.
- *
- * `before`/`edited` are the canvas pose just before and after the edit. Only
- * fields (and slots) that actually changed between them are merged: an effect
- * can own several fields (e.g. zoom owns both the main and per-slot scale), so
- * editing one must not overwrite a clip's keyframed value for the others. This
- * keeps `setScale` from propagating slot scale and `updateScreenshotSlot` from
- * propagating main fields or unrelated slots.
- */
-const mergeEffectsIntoPose = (
-  basePose: ClipBaseline,
-  before: ClipBaseline,
-  edited: ClipBaseline,
-  effects: AnimationEffect[]
-): ClipBaseline => {
-  const next: ClipBaseline = { ...basePose, slots: { ...basePose.slots } }
-  const nextRec = next as Record<string, unknown>
-  for (const effect of effects) {
-    for (const field of EFFECT_MAIN_POSE_FIELDS[effect]) {
-      if (poseValueEq(before[field], edited[field])) continue
-      nextRec[field] = edited[field]
-    }
-    const slotFields = EFFECT_SLOT_POSE_FIELDS[effect]
-    if (!slotFields) continue
-    for (const [slotId, editedSlot] of Object.entries(edited.slots)) {
-      const target = next.slots[slotId]
-      if (!target) continue
-      const beforeSlot = before.slots[slotId]
-      let mergedSlot = target
-      for (const field of slotFields) {
-        if (poseValueEq(beforeSlot?.[field], editedSlot[field])) continue
-        if (mergedSlot === target) mergedSlot = { ...target }
-        ;(mergedSlot as Record<string, unknown>)[field] = editedSlot[field]
-      }
-      if (mergedSlot !== target) next.slots[slotId] = mergedSlot
-    }
-  }
-  return next
-}
-
-/**
- * The resolved look AT a keyframe: for each effect, the value from the latest
- * keyframe that owns it at/before `target` (so held effects from earlier
- * keyframes show through), falling back to the final look for effects no
- * keyframe animates. Loading THIS onto the canvas makes selecting a keyframe show
- * its true accumulated state — which is what you edit.
- */
-const REST_SHADOW: Shadow = {
-  type: "none",
-  intensity: 0,
-  color: "#000000",
-  lightSource: "center",
-}
-
-// Non-bare (aspect-based) main-screenshot position geometry, mirroring
-// mainScreenshotPositionPct / mainScreenshotOffsetForPoint in position-math.ts
-// (which can't be imported here — it imports from this store). The razor split
-// uses these to interpolate position in the same percent-point space playback
-// uses, then invert the eased mid point back to a (cell, offset) keyframe.
-const POSITION_BASE_CANVAS_WIDTH = 1100
-const mainPositionBase = (
-  aspect: AspectState,
-  frame: DeviceFrame,
-  position: ScreenshotPosition,
-  slots: ScreenshotSlot[]
-) => {
-  const aw = aspect.w || 16
-  const ah = aspect.h || 10
-  const width = POSITION_BASE_CANVAS_WIDTH
-  const height = (POSITION_BASE_CANVAS_WIDTH * ah) / aw
-  const anchor = screenshotPositionAnchor(position)
-  let baseX = anchor.x
-  let baseY = anchor.y
-  // With extra screenshots, a centered main sits at its row-layout slot; other
-  // cells still anchor to the grid (matches mainScreenshotPositionPct).
-  if (slots.length > 0 && position === "center") {
-    const rowLayout = computeRowLayout(
-      [
-        { id: "__main__", frame },
-        ...slots.map((s) => ({ id: s.id, frame: s.frame ?? frame })),
-      ],
-      aw / ah
-    )
-    const mainLayout = rowLayout[0]
-    if (mainLayout) {
-      baseX = mainLayout.xPct
-      baseY = 50
-    }
-  }
-  return { width, height, baseX, baseY }
-}
-const mainPositionPoint = (
-  aspect: AspectState,
-  frame: DeviceFrame,
-  position: ScreenshotPosition,
-  offset: { x: number; y: number },
-  slots: ScreenshotSlot[]
-) => {
-  const { width, height, baseX, baseY } = mainPositionBase(
-    aspect,
-    frame,
-    position,
-    slots
-  )
-  return {
-    xPct: baseX + (offset.x / width) * 100,
-    yPct: baseY + (offset.y / height) * 100,
-  }
-}
-const mainPositionOffsetForPoint = (
-  aspect: AspectState,
-  frame: DeviceFrame,
-  position: ScreenshotPosition,
-  slots: ScreenshotSlot[],
-  point: { xPct: number; yPct: number }
-) => {
-  const { width, height, baseX, baseY } = mainPositionBase(
-    aspect,
-    frame,
-    position,
-    slots
-  )
-  return {
-    x: ((point.xPct - baseX) / 100) * width,
-    y: ((point.yPct - baseY) / 100) * height,
-  }
-}
-
-const resolveKeyframePose = (
-  canvas: CanvasState,
-  clips: AnimationClip[],
-  target: AnimationClip
-): ClipBaseline => {
-  const sorted = [...clips].sort((a, b) => a.startMs - b.startMs)
-  const last = sorted[sorted.length - 1]
-  const fallback = last ? clipPose(last) : captureClipPose(canvas)
-
-  // Latest keyframe that owns an effect at/before `target`, plus whether ANY
-  // keyframe owns it — so we can tell "held from an earlier keyframe" apart from
-  // "revealed by a later keyframe" (→ neutral rest) apart from "never animated"
-  // (→ the constant final value).
-  const latestOwner = (
-    owns: (c: AnimationClip) => boolean
-  ): { at: AnimationClip | null; any: boolean } => {
-    let at: AnimationClip | null = null
-    let any = false
-    for (const c of clips) {
-      if (!owns(c)) continue
-      any = true
-      if (c.startMs <= target.startMs && (!at || c.startMs >= at.startMs))
-        at = c
-    }
-    return { at, any }
-  }
-  const ownsMain = (effect: AnimationEffect) => (c: AnimationClip) =>
-    clipAffectsMain(c) && (c.effects ?? []).includes(effect)
-  const ownsSlot =
-    (slotId: string, effect: AnimationEffect) => (c: AnimationClip) =>
-      clipAffectsSlot(c, slotId) && (c.effects ?? []).includes(effect)
-
-  const main = <V,>(
-    effect: AnimationEffect,
-    extract: (p: ClipBaseline) => V,
-    rest: V
-  ): V => {
-    const { at, any } = latestOwner(ownsMain(effect))
-    if (at) return extract(clipPose(at))
-    return any ? rest : extract(fallback)
-  }
-
-  // Like `main`, but the "owned only by a LATER keyframe" case resolves to the
-  // value the effect REVEALS FROM — the first owner's baseline — instead of the
-  // final look. This mirrors playback (which eases from that baseline), so
-  // selecting a keyframe BEFORE the first change shows the pre-animation state,
-  // not the end state. Without this, editing a later keyframe (e.g. border →
-  // orange) makes every earlier keyframe resolve to that end value too.
-  const mainReveal = <V,>(
-    effect: AnimationEffect,
-    extract: (p: ClipBaseline) => V
-  ): V => {
-    const owners = clips
-      .filter((c) => ownsMain(effect)(c))
-      .sort((a, b) => a.startMs - b.startMs)
-    if (owners.length === 0) return extract(fallback)
-    const { at } = latestOwner(ownsMain(effect))
-    return extract(at ? clipPose(at) : clipBaseline(owners[0]))
-  }
-
-  return {
-    tilt: main("tilt", (p) => p.tilt, { rx: 0, ry: 0, rz: 0 }),
-    scale: main("zoom", (p) => p.scale, 100),
-    screenshotPosition: main("position", (p) => p.screenshotPosition, "center"),
-    screenshotOffset: main("position", (p) => p.screenshotOffset, {
-      x: 0,
-      y: 0,
-    }),
-    padding: main("padding", (p) => p.padding, 0),
-    canvasBorderRadius: main(
-      "canvasRadius",
-      (p) => p.canvasBorderRadius ?? canvas.canvasBorderRadius,
-      DEFAULT_BASELINE.canvasBorderRadius
-    ),
-    shadow: main("shadow", (p) => p.shadow, REST_SHADOW),
-    // Background is a single layer (no true crossfade), so always show the final.
-    background: (() => {
-      const { at } = latestOwner(ownsMain("background"))
-      return at ? clipPose(at).background : fallback.background
-    })(),
-    // Reveal effects: held from an earlier owner if one exists at/before this
-    // keyframe, else the value they reveal FROM (first owner's baseline) — never
-    // the final look, so an earlier keyframe never inherits a later edit.
-    filter: mainReveal("filter", (p) => p.filter ?? "none"),
-    mediaFilter: mainReveal(
-      "mediaFilter",
-      (p) => p.mediaFilter ?? canvas.mediaFilter ?? "none"
-    ),
-    mediaAdjustments: mainReveal(
-      "mediaEffects",
-      (p) =>
-        p.mediaAdjustments ??
-        canvas.mediaAdjustments ??
-        NEUTRAL_MEDIA_ADJUSTMENTS
-    ),
-    portrait: mainReveal("portrait", (p) => p.portrait ?? canvas.portrait),
-    pattern: mainReveal("pattern", (p) => p.pattern ?? canvas.backdrop.pattern),
-    overlay: mainReveal("overlay", (p) => p.overlay ?? canvas.overlay),
-    border: mainReveal("border", (p) => p.border ?? canvas.border),
-    crop: mainReveal("crop", (p) => poseCrop(p, canvas.lastCropRegion)),
-    borderRadius: main(
-      "borderRadius",
-      (p) => p.borderRadius ?? canvas.borderRadius,
-      DEFAULT_BASELINE.borderRadius ?? canvas.borderRadius
-    ),
-    backdropEffects: main(
-      "backdrop",
-      (p) => p.backdropEffects,
-      DEFAULT_BASELINE.backdropEffects
-    ),
-    lighting: main(
-      "lighting",
-      (p) => p.lighting ?? canvas.backdrop.lighting,
-      REST_LIGHTING
-    ),
-    slots: Object.fromEntries(
-      canvas.screenshotSlots.map((s) => {
-        const committed: ClipSlotPose = {
-          tilt: s.tilt,
-          scale: s.scale,
-          rotation: s.rotation,
-          shadow: s.shadow ?? canvas.shadow,
-          xPct: s.xPct,
-          yPct: s.yPct,
-        }
-        const slot = (effect: AnimationEffect, rest: ClipSlotPose) => {
-          const { at, any } = latestOwner(ownsSlot(s.id, effect))
-          if (at) return clipPose(at).slots[s.id] ?? rest
-          return any ? rest : committed
-        }
-        const t = slot("tilt", {
-          ...committed,
-          tilt: { rx: 0, ry: 0, rz: 0 },
-          rotation: 0,
-        })
-        const z = slot("zoom", { ...committed, scale: 100 })
-        const sh = slot("shadow", { ...committed, shadow: REST_SHADOW })
-        // Reveal-from semantics (mirrors mainReveal): held from the latest owner
-        // at/before this keyframe, else the value it reveals FROM (the first
-        // owner's captured baseline), else the committed slot value. Used for the
-        // effects that ease from the slot's own committed look — position, border,
-        // radius, padding, lighting — rather than from a neutral rest.
-        const slotReveal = <V,>(
-          effect: AnimationEffect,
-          extract: (sp: ClipSlotPose | undefined) => V | undefined,
-          committedVal: V
-        ): V => {
-          const owners = clips
-            .filter((c) => ownsSlot(s.id, effect)(c))
-            .sort((a, b) => a.startMs - b.startMs)
-          if (owners.length === 0) return committedVal
-          const { at } = latestOwner(ownsSlot(s.id, effect))
-          const src = at
-            ? clipPose(at).slots[s.id]
-            : clipBaseline(owners[0]).slots[s.id]
-          return extract(src) ?? committedVal
-        }
-        return [
-          s.id,
-          {
-            tilt: t.tilt,
-            rotation: t.rotation,
-            scale: z.scale,
-            shadow: sh.shadow ?? committed.shadow,
-            xPct: slotReveal("position", (sp) => sp?.xPct, s.xPct),
-            yPct: slotReveal("position", (sp) => sp?.yPct, s.yPct),
-            border: slotReveal(
-              "border",
-              (sp) => sp?.border,
-              s.border ?? canvas.border
-            ),
-            borderRadius: slotReveal(
-              "borderRadius",
-              (sp) => sp?.borderRadius,
-              s.borderRadius ?? canvas.borderRadius
-            ),
-            padding: slotReveal(
-              "padding",
-              (sp) => sp?.padding,
-              s.padding ?? canvas.padding
-            ),
-            lighting: slotReveal(
-              "lighting",
-              (sp) => sp?.lighting,
-              s.lighting ?? canvas.backdrop.lighting
-            ),
-            filter: slotReveal("mediaFilter", (sp) => sp?.filter, s.filter),
-            adjustments: slotReveal(
-              "mediaEffects",
-              (sp) => sp?.adjustments,
-              s.adjustments ?? canvas.mediaAdjustments
-            ),
-          },
-        ]
-      })
-    ),
-  }
-}
-
-/**
- * Which screenshot a newly-added clip should bind to, mirroring the inspector's
- * ScreenshotStyleTarget: a selected slot → that slot, else the main screenshot
- * if it's selected, else "all". Validates the slot still exists on `canvas`.
- */
-const resolveSelectionTarget = (
-  canvas: CanvasState,
-  selectedScreenshotSlotId: string | null,
-  isScreenshotSelected: boolean
-): AnimationClipTarget => {
-  if (
-    selectedScreenshotSlotId &&
-    canvas.screenshotSlots.some((s) => s.id === selectedScreenshotSlotId)
-  ) {
-    return { scope: "slot", slotId: selectedScreenshotSlotId }
-  }
-  if (isScreenshotSelected) return { scope: "main" }
-  return { scope: "all" }
-}
+import { createAnimationActions } from "./store/actions/animation"
+import { createCanvasStyleActions } from "./store/actions/canvas-style"
+import { createCanvasActions } from "./store/actions/canvases"
+import { createLayerActions } from "./store/actions/layers"
+import { createMediaActions } from "./store/actions/media"
+import { createProjectActions } from "./store/actions/project"
+import { createSessionActions } from "./store/actions/session"
+import { createSlotActions } from "./store/actions/slots"
+import { createCommitContext } from "./store/commit-context"
+import { INITIAL_EDITOR_STORE_STATE } from "./store/initial-state"
+import type { EditorStore } from "./store/types"
 
 export * from "./state-types"
 export {
@@ -803,8 +52,10 @@ export {
   sampleImageColorsRaw,
 } from "./color-utils"
 
-export { MAX_CANVASES, MAX_SCREENSHOT_SLOTS }
-export type { PresetTab, CurrentDraftInfo }
+export { MAX_CANVASES, MAX_SCREENSHOT_SLOTS } from "./store/defaults"
+export { captureClipPose } from "./store/animation-helpers"
+export type { PresetTab } from "./store/canvas-helpers"
+export type { CurrentDraftInfo } from "./store/draft-persistence"
 
 export {
   CanvasPreviewScope,
@@ -825,3254 +76,31 @@ export {
   saveEditorDraftBeforeAuth,
 } from "./store/provider"
 
-type SetPatch =
-  | Partial<EditorState>
-  | ((state: EditorState) => Partial<EditorState>)
-
-type CanvasPatch =
-  | Partial<CanvasState>
-  | ((canvas: CanvasState, state: EditorState) => Partial<CanvasState>)
-
-export type CustomPresetSlotConfig = {
-  xPct: number
-  yPct: number
-  widthPct?: number
-  heightPct?: number
-  rotation: number
-  tilt: Tilt
-  scale: number
-  zIndex?: number
-  filter?: AssetFilter
-  adjustments?: MediaAdjustments
-  hidden?: boolean
-  objectFit?: "contain" | "cover" | "fill"
-  shadow?: Shadow
-}
-
-/**
- * Full visual snapshot of a canvas that gets re-applied when a custom preset is
- * selected — every styling field but never the actual screenshot pixels. The
- * exact field set is derived from {@link PRESET_NON_STYLE_KEYS}: capture, apply,
- * and preview all read it from `preset-fields` so a new canvas styling field
- * flows into presets with no change here.
- */
-export type { CustomPresetCanvasStyle }
-
-/** User-saved custom preset kind: static look vs timeline (animate). */
-export type CustomPresetType = "style" | "animate"
-
-/**
- * Timeline payload for animate presets. Audio is never stored; clip/slot ids
- * are remapped on apply via `sourceSlotIds` (geometry.slots order).
- */
-export type CustomPresetAnimation = {
-  durationMs: number
-  clips: AnimationClip[]
-  /** Slot ids from the source canvas at save time, in geometry.slots order. */
-  sourceSlotIds?: string[]
-}
-
-export type CustomPresetGeometry = {
-  canvasTilt: Tilt
-  canvasScale: number
-  slots: CustomPresetSlotConfig[]
-  mainOffset?: { xPct: number; yPct: number }
-  relativeSlotPositions?: boolean
-  canvasStyle?: CustomPresetCanvasStyle
-  animation?: CustomPresetAnimation
-}
-
-export type CustomPresetSummary = {
-  id: string
-  name: string
-  slotCount: number
-  /** Defaults to "style" for older presets that predate the column. */
-  type?: CustomPresetType
-  geometry: CustomPresetGeometry
-}
-
-export type DraftLoadUi = {
-  presetTab?: PresetTab
-  activeLayoutPresetId?: string | null
-  activeCustomPresetId?: string | null
-  activeSinglePresetId?: string | null
-  bulkEditMode?: boolean
-  bulkViewportZoom?: number
-  bulkScale?: number
-  previewAutoScrollDelay?: number
-  previewAnimation?: "slide" | "fade" | "zoom" | "flip"
-  /** When true, re-enter Animate mode after loading the project. */
-  isAnimateMode?: boolean
-}
-
-export type EditorActions = {
-  setTopBarPopoverOpen: (open: boolean) => void
-  setActiveTool: (t: EditorTool) => void
-  setPresetTab: (tab: PresetTab) => void
-  setActiveLayoutPresetId: (id: string | null) => void
-  setActiveCustomPresetId: (id: string | null) => void
-  setActiveSinglePresetId: (id: string | null) => void
-  setCustomPresets: (presets: CustomPresetSummary[]) => void
-  /** Logout / session clear — empties the list without marking a successful load. */
-  clearCustomPresets: () => void
-  /**
-   * Fetch the signed-in user's custom presets once. Dedupes in-flight calls so
-   * multiple mounted Preset sections (desktop + iPad sidebars) share one request.
-   */
-  loadCustomPresets: (userId: string, sort?: PresetSort) => void
-  addCustomPreset: (preset: CustomPresetSummary) => void
-  updateCustomPreset: (id: string, patch: Partial<CustomPresetSummary>) => void
-  removeCustomPreset: (id: string) => void
-  setCurrentDraft: (draft: CurrentDraftInfo | null) => void
-  loadDraftState: (
-    state: Partial<EditorState>,
-    draft: CurrentDraftInfo,
-    ui?: DraftLoadUi
-  ) => void
-  /**
-   * Apply a bundled template's full composition as a brand-new, unsaved
-   * project. Same restore path as {@link loadDraftState} but never links to a
-   * saved draft, so the next Save creates a fresh draft.
-   */
-  loadTemplateState: (state: Partial<EditorState>, ui?: DraftLoadUi) => void
-  applyPresetSnapshot: (
-    snapshot: CustomPresetGeometry,
-    canvasId?: string
-  ) => void
-  setScreenshot: (s: string | null, canvasId?: string) => void
-  setFullPageScreenshot: (src: string | null, canvasId?: string) => void
-  setFullPageScreenshotScrollPosition: (
-    scrollPosition: number,
-    canvasId?: string
-  ) => void
-  applyCroppedScreenshot: (
-    s: string,
-    region: CropRegion,
-    canvasId?: string
-  ) => void
-  // Non-destructive crop: stores a render-time crop region without re-encoding
-  // or replacing the screenshot src. Used for video, which can't be re-encoded
-  // client-side; passing null clears the crop.
-  setScreenshotCropRegion: (
-    region: CropRegion | null,
-    canvasId?: string
-  ) => void
-  /** Update the placement or non-destructive in/out range of one video section. */
-  updateVideoClip: (
-    id: string,
-    patch: Partial<Omit<VideoTimelineClip, "id">>,
-    canvasId?: string
-  ) => void
-  /** Split a video section at a source-timeline time. */
-  splitVideoClip: (id: string, atMs: number, canvasId?: string) => string | null
-  /** Copy a video section and insert it after its source range. */
-  duplicateVideoClip: (
-    id: string,
-    durationMs: number,
-    canvasId?: string
-  ) => string | null
-  /** Delete one or more video sections. */
-  removeVideoClips: (ids: string[], canvasId?: string) => void
-  setAspect: (a: AspectState) => void
-  setCanvasAspect: (canvasId: string, a: AspectState) => void
-  setBackground: (
-    b: Background,
-    canvasId?: string,
-    opts?: { silent?: boolean }
-  ) => void
-  setPadding: (n: number, canvasId?: string) => void
-  setBorderRadius: (n: number, canvasId?: string) => void
-  setCanvasBorderRadius: (n: number, canvasId?: string) => void
-  setBorder: (b: Border, canvasId?: string) => void
-  setMainScreenshotPadding: (n: number, canvasId?: string) => void
-  setMainScreenshotBorderRadius: (n: number, canvasId?: string) => void
-  setMainScreenshotBorder: (b: Border, canvasId?: string) => void
-  setBackdropEffects: (e: BackdropEffects, canvasId?: string) => void
-  setBackdropPattern: (p: BackdropPattern, canvasId?: string) => void
-  setBackdropLighting: (l: BackdropLighting, canvasId?: string) => void
-  setMainScreenshotBackdropLighting: (
-    l: BackdropLighting,
-    canvasId?: string
-  ) => void
-  setBackdropFilter: (f: AssetFilter, canvasId?: string) => void
-  setTilt: (t: Tilt, canvasId?: string) => void
-  setScale: (n: number, canvasId?: string) => void
-  setTiltAndScale: (t: Tilt, scale: number, canvasId?: string) => void
-  setScreenshotTilt: (t: Tilt, canvasId?: string) => void
-  setScreenshotScale: (n: number, canvasId?: string) => void
-  setScreenshotRotation: (n: number, canvasId?: string) => void
-  setCanvasZoom: (n: number) => void
-  setScreenshotPosition: (p: ScreenshotPosition, canvasId?: string) => void
-  setScreenshotOffset: (o: { x: number; y: number }, canvasId?: string) => void
-  setScreenshotPlacement: (
-    p: ScreenshotPosition,
-    o: { x: number; y: number },
-    canvasId?: string
-  ) => void
-  updateScreenshotLayer: (
-    patch: Partial<ScreenshotLayer>,
-    canvasId?: string
-  ) => void
-  setShadow: (s: Shadow, canvasId?: string) => void
-  setMainScreenshotShadow: (s: Shadow, canvasId?: string) => void
-  /**
-   * Single entry point for screenshot style edits. `target` decides whether the
-   * patch lands on the main screenshot, a specific slot, or all of them; the
-   * inspector uses this instead of picking between main/all/slot setters.
-   */
-  applyScreenshotStyle: (
-    target: ScreenshotStyleTarget,
-    patch: ScreenshotStylePatch,
-    canvasId?: string
-  ) => void
-  setOverlay: (o: Overlay, canvasId?: string) => void
-  setFrame: (f: DeviceFrame, canvasId?: string) => void
-  setFrameForMatchingScreenshots: (f: DeviceFrame, canvasId?: string) => void
-  setMainScreenshotFrame: (f: DeviceFrame, canvasId?: string) => void
-  setFrameAddress: (address: string, canvasId?: string) => void
-  setTweet: (card: TweetCard, canvasId?: string) => void
-  updateTweet: (patch: Partial<TweetCard>, canvasId?: string) => void
-  clearTweet: (canvasId?: string) => void
-  setObjectFit: (fit: "contain" | "cover" | "fill", canvasId?: string) => void
-  bringScreenshotToFront: (canvasId?: string) => void
-  sendScreenshotToBack: (canvasId?: string) => void
-  setPortrait: (p: Portrait, canvasId?: string) => void
-  setEnhance: (e: EnhancePreset, canvasId?: string) => void
-  setAnnotation: (patch: Partial<Annotation>) => void
-  addAnnotationStroke: (
-    stroke: Omit<AnnotationStroke, "id" | "zIndex">,
-    canvasId?: string
-  ) => string
-  updateAnnotationStroke: (
-    id: string,
-    points: AnnotationPoint[],
-    canvasId?: string
-  ) => void
-  updateAnnotationStrokeLayer: (
-    id: string,
-    patch: Partial<
-      Pick<AnnotationStroke, "zIndex" | "opacity" | "blendMode" | "hidden">
-    >,
-    canvasId?: string
-  ) => void
-  deleteAnnotationStroke: (id: string, canvasId?: string) => void
-  addAnnotationShape: (
-    shape: Omit<AnnotationShape, "id" | "zIndex">,
-    canvasId?: string
-  ) => string
-  updateAnnotationShape: (
-    id: string,
-    patch: Partial<AnnotationShape>,
-    canvasId?: string
-  ) => void
-  deleteAnnotationShape: (id: string, canvasId?: string) => void
-  duplicateAnnotationShape: (id: string, canvasId?: string) => string | null
-  bringAnnotationShapeToFront: (id: string, canvasId?: string) => void
-  sendAnnotationShapeToBack: (id: string, canvasId?: string) => void
-  clearAnnotations: (canvasId?: string) => void
-  addText: (canvasId?: string) => string
-  updateText: (
-    id: string,
-    patch: Partial<TextElement>,
-    canvasId?: string
-  ) => void
-  deleteText: (id: string, canvasId?: string) => void
-  duplicateText: (id: string, canvasId?: string) => string | null
-  bringTextToFront: (id: string, canvasId?: string) => void
-  sendTextToBack: (id: string, canvasId?: string) => void
-  setSelectedTextId: (id: string | null) => void
-  addAsset: (src: string, canvasId?: string) => string
-  updateAsset: (
-    id: string,
-    patch: Partial<AssetElement>,
-    canvasId?: string
-  ) => void
-  deleteAsset: (id: string, canvasId?: string) => void
-  duplicateAsset: (id: string, canvasId?: string) => string | null
-  bringAssetToFront: (id: string, canvasId?: string) => void
-  sendAssetToBack: (id: string, canvasId?: string) => void
-  setSelectedAssetId: (id: string | null) => void
-  setSelectedAnnotationShapeId: (id: string | null) => void
-  setSelectedScreenshotSlotId: (id: string | null) => void
-  setIsScreenshotSelected: (selected: boolean) => void
-  setIsAnimateMode: (a: boolean) => void
-  /**
-   * Open a clip for editing: saves the currently-open clip's pose from the live
-   * canvas, then loads the newly-selected clip's pose onto the canvas so the
-   * inspector edits that clip's keyframe. Pass null to deselect.
-   */
-  selectAnimationClip: (id: string | null, canvasId?: string) => void
-  setAnimationDuration: (ms: number, canvasId?: string) => void
-  addAnimationClip: (canvasId?: string, atMs?: number) => string
-  updateAnimationClip: (
-    id: string,
-    patch: Partial<Omit<AnimationClip, "id">>,
-    canvasId?: string
-  ) => void
-  /**
-   * Strip every animated effect from a clip: reverts its pose to its captured
-   * baseline and clears `effects` so it animates nothing. When it's the open
-   * clip, the committed canvas reverts too (so e.g. a lit backdrop goes dark).
-   */
-  clearAnimationClipEffects: (id: string, canvasId?: string) => void
-  removeAnimationClip: (id: string, canvasId?: string) => void
-  moveAnimationClip: (id: string, startMs: number, canvasId?: string) => void
-  duplicateAnimationClip: (id: string, canvasId?: string) => string | null
-  /**
-   * Replace the timeline selection with `ids` (a marquee drag). One id behaves
-   * exactly like `selectAnimationClip`; several select a group for bulk actions
-   * and open no single keyframe for editing.
-   */
-  setAnimationClipSelection: (ids: string[], canvasId?: string) => void
-  /** Bulk-delete every clip in `ids` in a single history entry. */
-  removeAnimationClips: (ids: string[], canvasId?: string) => void
-  /** Bulk "remove effects" — strip animated effects from every clip in `ids`. */
-  clearAnimationClipsEffects: (ids: string[], canvasId?: string) => void
-  /** Bulk-duplicate every clip in `ids`; returns the new clip ids. */
-  duplicateAnimationClips: (ids: string[], canvasId?: string) => string[]
-  /**
-   * Cut a clip in two at `atMs` (like a razor tool). The first half keeps the
-   * original id/pose; the second half is a new clip holding the same target
-   * keyframe. Together they fill the original clip's exact footprint, so each
-   * can then be dragged/trimmed on its own. Returns the new (second) clip id, or
-   * null if the cut leaves either side smaller than the minimum clip length.
-   */
-  splitAnimationClip: (
-    id: string,
-    atMs: number,
-    canvasId?: string
-  ) => string | null
-  clearAnimationClips: (canvasId?: string) => void
-  setIsPreviewMode: (p: boolean) => void
-  setIsPreviewAutoScroll: (a: boolean) => void
-  setPreviewAutoScrollDelay: (d: number) => void
-  setPreviewAnimation: (a: "slide" | "fade" | "zoom" | "flip") => void
-  setBulkEditMode: (b: boolean) => void
-  setBulkCanvasDragging: (dragging: boolean) => void
-  setScreenshotPositionDragging: (dragging: boolean) => void
-  setBulkViewportZoom: (zoom: number) => void
-  setBulkScale: (n: number) => void
-  reset: () => void
-  undo: () => void
-  redo: () => void
-  addCanvas: () => string | null
-  removeCanvas: (id: string) => void
-  duplicateCanvas: (id?: string) => string | null
-  setActiveCanvasId: (id: string) => void
-  setCanvasPosition: (id: string, position: { x: number; y: number }) => void
-  setCanvasPositions: (
-    positions: Record<string, { x: number; y: number }>
-  ) => void
-  requestBulkFitView: () => void
-  addScreenshotSlot: (canvasId?: string) => string | null
-  updateScreenshotSlot: (
-    id: string,
-    patch: Partial<ScreenshotSlot>,
-    canvasId?: string
-  ) => void
-  setScreenshotSlotImage: (
-    id: string,
-    src: string | null,
-    canvasId?: string
-  ) => void
-  setFullPageScreenshotSlot: (
-    id: string,
-    src: string | null,
-    canvasId?: string
-  ) => void
-  setFullPageScreenshotSlotScrollPosition: (
-    id: string,
-    scrollPosition: number,
-    canvasId?: string
-  ) => void
-  applyCroppedScreenshotSlot: (
-    id: string,
-    src: string,
-    region: CropRegion,
-    canvasId?: string
-  ) => void
-  deleteScreenshotSlot: (id: string, canvasId?: string) => void
-  duplicateScreenshotSlot: (id: string, canvasId?: string) => string | null
-  bringScreenshotSlotToFront: (id: string, canvasId?: string) => void
-  sendScreenshotSlotToBack: (id: string, canvasId?: string) => void
-  arrangeScreenshotSlotsInRow: (canvasId?: string) => void
-  setScreenshotSlotGroupPosition: (
-    position: { xPct: number; yPct: number },
-    canvasId?: string
-  ) => void
-}
-
-export type EditorStore = {
-  past: EditorState[]
-  present: EditorState
-  future: EditorState[]
-  _lastGroup: string | null
-  _lastTs: number
-  topBarPopoverOpen: boolean
-  isAnimateMode: boolean
-  isPreviewMode: boolean
-  isPreviewAutoScroll: boolean
-  previewAutoScrollDelay: number
-  previewAnimation: "slide" | "fade" | "zoom" | "flip"
-  bulkEditMode: boolean
-  bulkCanvasDragging: boolean
-  /** True while a screenshot/slot is being moved via the position pad (not the
-   * on-canvas box drag). Lets the boxes drop their left/top/transform easing so
-   * they track the pad live instead of easing ~300ms behind it. */
-  screenshotPositionDragging: boolean
-  bulkViewportZoom: number
-  bulkScale: number
-  bulkFitViewSeq: number
-  selectedTextId: string | null
-  selectedAssetId: string | null
-  selectedAnnotationShapeId: string | null
-  selectedScreenshotSlotId: string | null
-  isScreenshotSelected: boolean
-  /** Timeline clip currently open for editing in Animate mode (its keyframe). */
-  selectedAnimationClipId: string | null
-  /**
-   * Every clip selected in the timeline — the set bulk context-menu / keyboard
-   * actions operate on. Contains `selectedAnimationClipId` when a single clip is
-   * open; a marquee drag can select several (then no single clip is "open" for
-   * editing, so `selectedAnimationClipId` is null).
-   */
-  selectedAnimationClipIds: string[]
-  presetTab: PresetTab
-  activeLayoutPresetId: string | null
-  activeCustomPresetId: string | null
-  activeSinglePresetId: string | null
-  customPresets: CustomPresetSummary[]
-  customPresetsLoaded: boolean
-  customPresetsLoading: boolean
-  /**
-   * Set when the last load failed. Without it a failed fetch is indistinguishable
-   * from an empty account, and the picker tells the user they have no presets
-   * saved while their presets are sitting on the server.
-   */
-  customPresetsError: boolean
-  /** User id the current `customPresets` list was fetched for. */
-  customPresetsForUserId: string | null
-  /**
-   * Sort currently requested — drives the sort dropdown. Leads the displayed
-   * list while a re-sort request is in flight.
-   */
-  customPresetsSort: PresetSort
-  /**
-   * Sort the displayed `customPresets` list is actually ordered by. Updated only
-   * once a sort response lands, so `addCustomPreset` and failure rollback use the
-   * order the user is really looking at rather than a pending request.
-   */
-  customPresetsListSort: PresetSort
-  currentDraft: CurrentDraftInfo | null
-} & EditorActions
-
-/**
- * Insert a copy of `sourceId` right after it in the array, rippling later clips
- * right just enough to open a gap (preserving their spacing). Shared by single-
- * and bulk-duplicate.
- */
-const insertClipCopy = (
-  clips: AnimationClip[],
-  sourceId: string,
-  newId: string
-): AnimationClip[] => {
-  const source = clips.find((c) => c.id === sourceId)
-  if (!source) return clips
-  const dur = source.durationMs
-  const insertStart = Math.min(
-    source.startMs + dur,
-    Math.max(0, MAX_DURATION_MS - dur)
-  )
-  const nextStart = clips
-    .filter((clip) => clip.id !== source.id && clip.startMs >= insertStart)
-    .reduce((min, clip) => Math.min(min, clip.startMs), Infinity)
-  const shift = Number.isFinite(nextStart)
-    ? Math.max(0, insertStart + dur - nextStart)
-    : 0
-  const shifted = clips.map((clip) =>
-    clip.id !== source.id && clip.startMs >= insertStart
-      ? {
-          ...clip,
-          startMs: Math.min(
-            clip.startMs + shift,
-            Math.max(0, MAX_DURATION_MS - clip.durationMs)
-          ),
-        }
-      : clip
-  )
-  const copy: AnimationClip = { ...source, id: newId, startMs: insertStart }
-  const sourceIndex = shifted.findIndex((cl) => cl.id === source.id)
-  return [
-    ...shifted.slice(0, sourceIndex + 1),
-    copy,
-    ...shifted.slice(sourceIndex + 1),
-  ]
-}
-
-/**
- * Strip a clip's animated effects in-array: reverts its pose to its captured
- * baseline and clears `effects`. Returns the same array reference when the clip
- * owns nothing (so callers can detect a no-op). Shared by single- and bulk-clear.
- */
-const clearClipEffectsInArray = (
-  clips: AnimationClip[],
-  id: string
-): AnimationClip[] => {
-  const clip = clips.find((c) => c.id === id)
-  if (!clip || (clip.effects ?? []).length === 0) return clips
-  const cleared: AnimationClip = {
-    ...clip,
-    effects: [],
-    pose: { ...DEFAULT_BASELINE, ...clipBaseline(clip) },
-  }
-  return clips.map((c) => (c.id === id ? cleared : c))
-}
-
-// Identity of the newest custom-presets request. A response that doesn't match
-// is stale — cleared/logged out, or superseded by a load for another account —
-// and must not commit results or touch the loading flag.
-let customPresetsRequestToken = 0
-let customPresetsInFlightUserId: string | null = null
-
+export type {
+  CustomPresetAnimation,
+  CustomPresetCanvasStyle,
+  CustomPresetGeometry,
+  CustomPresetSlotConfig,
+  CustomPresetSummary,
+  CustomPresetType,
+  DraftLoadUi,
+  EditorActions,
+  EditorStore,
+  EditorStoreState,
+} from "./store/types"
 export const useEditorStore = create<EditorStore>((set, get) => {
-  const commit = (patch: SetPatch, group: string | null) => {
-    const state = get()
-    const resolvedPatch =
-      typeof patch === "function" ? patch(state.present) : patch
-    const present = { ...state.present, ...resolvedPatch }
-    const now = Date.now()
-    const canMerge =
-      group !== null &&
-      group === state._lastGroup &&
-      now - state._lastTs < GROUP_MERGE_MS
-    if (canMerge) {
-      set({ present, future: [], _lastTs: now })
-      return
-    }
-    const past = [...state.past, state.present]
-    if (past.length > HISTORY_LIMIT) past.shift()
-    set({
-      past,
-      present,
-      future: [],
-      _lastGroup: group,
-      _lastTs: now,
-    })
-  }
-
-  const commitCanvas = (
-    targetId: string | undefined,
-    patch: CanvasPatch,
-    group: string | null
-  ) => {
-    commit((state) => {
-      const canvasId = targetId ?? state.activeCanvasId
-      const canvases = state.canvases.map((canvas) => {
-        if (canvas.id !== canvasId) return canvas
-        const resolvedPatch =
-          typeof patch === "function" ? patch(canvas, state) : patch
-        return { ...canvas, ...resolvedPatch }
-      })
-      return { canvases }
-    }, group)
-  }
-
-  /**
-   * Like `commitCanvas`, but while a keyframe is open in Animate mode it also
-   * records `effects` as owned by that keyframe. This is how "changing an effect
-   * while a clip is selected makes that clip own it" — the ONLY thing a keyframe
-   * animates is the effect set recorded here.
-   */
-  const commitCanvasEffect = (
-    targetId: string | undefined,
-    patch: CanvasPatch,
-    group: string | null,
-    effects: AnimationEffect | AnimationEffect[]
-  ) => {
-    const list = Array.isArray(effects) ? effects : [effects]
-    commitCanvas(
-      targetId,
-      (canvas, state) => {
-        const base = typeof patch === "function" ? patch(canvas, state) : patch
-        const full = get()
-        // A patch that touches nothing animatable must not reach the keyframe
-        // bookkeeping below: an empty list satisfies every `list.every(...)`
-        // test and would re-target an unbound clip on a non-animatable edit.
-        if (!full.isAnimateMode || list.length === 0) return base
-        const anim = getCanvasAnimation(canvas)
-        const selId = full.selectedAnimationClipId
-
-        // Multi-selection opens no single keyframe (primary = null), so route the
-        // edit to EVERY selected clip: record the effect on each and write its new
-        // value into each clip's pose, leaving their other keyframed values intact.
-        const multiIds = full.selectedAnimationClipIds
-        if (!selId && multiIds.length > 1) {
-          const idSet = new Set(multiIds)
-          if (!anim.clips.some((c) => idSet.has(c.id))) return base
-          const beforePose = captureClipPose(canvas)
-          const editedPose = captureClipPose({
-            ...canvas,
-            ...base,
-          })
-          return {
-            ...base,
-            animation: {
-              ...anim,
-              clips: anim.clips.map((c) => {
-                if (!idSet.has(c.id)) return c
-                const owned = c.effects ?? []
-                const merged = Array.from(new Set([...owned, ...list]))
-                const basePose =
-                  c.pose ?? resolveKeyframePose(canvas, anim.clips, c)
-                return {
-                  ...c,
-                  effects: merged,
-                  pose: mergeEffectsIntoPose(
-                    basePose,
-                    beforePose,
-                    editedPose,
-                    list
-                  ),
-                }
-              }),
-            },
-          }
-        }
-
-        if (!selId) return base
-        const clip = anim.clips.find((c) => c.id === selId)
-        if (!clip) return base
-        const owned = clip.effects ?? []
-        const merged = Array.from(new Set([...owned, ...list]))
-        // Auto-bind: an as-yet-unbound ("all") keyframe binds to the SLOT this
-        // edit targets, so selecting it later re-selects that slot and further
-        // edits scope to it. Only for the effects a slot can actually animate
-        // (SLOT_ANIMATABLE_EFFECTS: transform, shadow, position, border, radius,
-        // padding, lighting) — canvas-wide effects (background, backdrop, pattern,
-        // filter, portrait, overlay, canvas radius) always edit the canvas/main
-        // even when a slot is selected, so they must not re-bind the clip. A
-        // keyframe already bound keeps its binding.
-        const currentTarget = clip.target ?? { scope: "all" as const }
-        const nextTarget = resolveSelectionTarget(
-          canvas,
-          full.selectedScreenshotSlotId,
-          full.isScreenshotSelected
-        )
-        const retarget =
-          currentTarget.scope === "all" &&
-          nextTarget.scope === "slot" &&
-          list.every((e) => SLOT_ANIMATABLE_EFFECTS.includes(e))
-        if (list.every((e) => owned.includes(e)) && !retarget) return base
-        return {
-          ...base,
-          animation: {
-            ...anim,
-            clips: anim.clips.map((c) =>
-              c.id === selId
-                ? {
-                    ...c,
-                    effects: merged,
-                    ...(retarget ? { target: nextTarget } : {}),
-                  }
-                : c
-            ),
-          },
-        }
-      },
-      group
-    )
-  }
-
-  const makeLayerOps = (
-    prefix: string,
-    getGroup?: (id: string) => string | null
-  ) => ({
-    toFront: (id: string, canvasId?: string) =>
-      commitCanvas(
-        canvasId,
-        (c) => moveLayerInStack(c, `${prefix}:${id}`, "front"),
-        getGroup?.(id) ?? null
-      ),
-    toBack: (id: string, canvasId?: string) =>
-      commitCanvas(
-        canvasId,
-        (c) => moveLayerInStack(c, `${prefix}:${id}`, "back"),
-        getGroup?.(id) ?? null
-      ),
-  })
-
-  const textLayerOps = makeLayerOps("text")
-  const assetLayerOps = makeLayerOps("asset")
-  const annotationShapeLayerOps = makeLayerOps("annotation")
-  const slotLayerOps = makeLayerOps("slot", (id) => `slot-layer-${id}`)
+  const context = createCommitContext(set, get)
 
   return {
-    past: [],
-    present: DEFAULT_STATE,
-    future: [],
-    _lastGroup: null,
-    _lastTs: 0,
-    topBarPopoverOpen: false,
-    isAnimateMode: false,
-    isPreviewMode: false,
-    isPreviewAutoScroll: false,
-    previewAutoScrollDelay: 3000,
-    previewAnimation: "slide" as const,
-    bulkEditMode: false,
-    bulkCanvasDragging: false,
-    screenshotPositionDragging: false,
-    bulkViewportZoom: 1,
-    bulkScale: 65,
-    bulkFitViewSeq: 0,
-    selectedTextId: null,
-    selectedAssetId: null,
-    selectedAnnotationShapeId: null,
-    selectedScreenshotSlotId: null,
-    isScreenshotSelected: false,
-    selectedAnimationClipId: null,
-    selectedAnimationClipIds: [],
-    presetTab: "single",
-    activeLayoutPresetId: null,
-    activeCustomPresetId: null,
-    activeSinglePresetId: null,
-    customPresets: [],
-    customPresetsLoaded: false,
-    customPresetsLoading: false,
-    customPresetsError: false,
-    customPresetsForUserId: null,
-    customPresetsSort: "latest",
-    customPresetsListSort: "latest",
-    currentDraft: null,
+    ...INITIAL_EDITOR_STORE_STATE,
 
-    setActiveTool: (t) => commit({ activeTool: t }, null),
-    setPresetTab: (tab) => set({ presetTab: tab }),
-    setActiveLayoutPresetId: (id) => set({ activeLayoutPresetId: id }),
-    setActiveCustomPresetId: (id) => set({ activeCustomPresetId: id }),
-    setActiveSinglePresetId: (id) => set({ activeSinglePresetId: id }),
-    setCustomPresets: (presets) =>
-      set({
-        customPresets: presets,
-        customPresetsLoaded: true,
-        customPresetsError: false,
-      }),
-    clearCustomPresets: () => {
-      customPresetsRequestToken++
-      customPresetsInFlightUserId = null
-      set({
-        customPresets: [],
-        customPresetsLoaded: false,
-        customPresetsLoading: false,
-        customPresetsError: false,
-        customPresetsForUserId: null,
-      })
-    },
-    loadCustomPresets: (userId, sort) => {
-      const state = get()
-      const nextSort = sort ?? state.customPresetsSort
-      const sortChanged = nextSort !== state.customPresetsSort
-      // A failed load also lands on `customPresetsLoaded` (it stops the
-      // skeleton), so it must not dedupe away the retry.
-      if (
-        state.customPresetsLoaded &&
-        !state.customPresetsError &&
-        state.customPresetsForUserId === userId &&
-        !sortChanged
-      )
-        return
-      // Dedupe concurrent mounts only for the SAME account. A load for a
-      // different user supersedes the in-flight request instead of being
-      // blocked by it — its late response is dropped by the token check.
-      // A sort change also supersedes: the in-flight list is ordered wrong.
-      if (
-        state.customPresetsLoading &&
-        customPresetsInFlightUserId === userId &&
-        !sortChanged
-      )
-        return
-      const token = ++customPresetsRequestToken
-      customPresetsInFlightUserId = userId
-      set({
-        customPresetsLoading: true,
-        customPresetsError: false,
-        customPresetsSort: nextSort,
-        // Account switch: never show the previous account's presets while the
-        // new list loads. A sort change keeps the current list on screen and
-        // reorders it in place when the response lands.
-        ...(state.customPresetsForUserId !== null &&
-        state.customPresetsForUserId !== userId
-          ? {
-              customPresets: [],
-              customPresetsLoaded: false,
-              customPresetsForUserId: null,
-            }
-          : {}),
-      })
-      void fetch(`/api/presets?sort=${nextSort}`, { credentials: "include" })
-        .then(async (res) => {
-          if (!res.ok) throw new Error(`Preset load failed: ${res.status}`)
-          const body: { presets: CustomPresetSummary[] } = await res.json()
-          return body.presets
-        })
-        .then((presets) => {
-          // Stale: cleared/logged out, or another account's load took over.
-          if (token !== customPresetsRequestToken) return
-          customPresetsInFlightUserId = null
-          set({
-            customPresets: presets,
-            customPresetsLoaded: true,
-            customPresetsLoading: false,
-            customPresetsError: false,
-            customPresetsForUserId: userId,
-            // The displayed list now reflects the requested sort.
-            customPresetsListSort: nextSort,
-          })
-        })
-        .catch((err) => {
-          console.error("Could not load custom presets", err)
-          if (token !== customPresetsRequestToken) return
-          customPresetsInFlightUserId = null
-          // Keep whatever list is on screen: a failed re-sort must not erase
-          // presets that loaded fine. Roll the requested sort back to the order
-          // actually displayed so the picker matches it.
-          set((current) => {
-            const rolledSort = current.customPresetsLoaded
-              ? current.customPresetsListSort
-              : nextSort
-            return {
-              customPresetsSort: rolledSort,
-              customPresetsListSort: rolledSort,
-              customPresetsLoaded: true,
-              customPresetsLoading: false,
-              // Only an empty list is a lie about the account. A failed re-sort
-              // still has real presets on screen, so it stays a plain list.
-              customPresetsError: current.customPresets.length === 0,
-              customPresetsForUserId: userId,
-            }
-          })
-        })
-    },
-    addCustomPreset: (preset) =>
-      set((state) => ({
-        // A new preset is the newest, so it belongs at whichever end the
-        // displayed order puts newest — otherwise saving under "Oldest" drops it
-        // at the top, where a refetch would not have put it. Uses the list's real
-        // sort, not a sort request still in flight.
-        customPresets:
-          state.customPresetsListSort === "oldest"
-            ? [...state.customPresets, preset]
-            : [preset, ...state.customPresets],
-        customPresetsLoaded: true,
-      })),
-    updateCustomPreset: (id, patch) =>
-      set((state) => ({
-        customPresets: state.customPresets.map((p) =>
-          p.id === id ? { ...p, ...patch } : p
-        ),
-      })),
-    removeCustomPreset: (id) =>
-      set((state) => ({
-        customPresets: state.customPresets.filter((p) => p.id !== id),
-        activeCustomPresetId:
-          state.activeCustomPresetId === id ? null : state.activeCustomPresetId,
-      })),
-    setCurrentDraft: (draft) => set({ currentDraft: draft }),
-    loadDraftState: (state, draft, ui) => {
-      const present = normalizeEditorState(state)
-      const defaultBulk = present.canvases.length > 1
-      const restoreAnimate = Boolean(ui?.isAnimateMode)
-      // When re-entering Animate, select the last clip on the active canvas so
-      // the timeline opens ready to edit (matches setIsAnimateMode(true)).
-      let selectedClipId: string | null = null
-      if (restoreAnimate) {
-        const active = present.canvases.find(
-          (c) => c.id === present.activeCanvasId
-        )
-        const clips = active?.animation?.clips ?? []
-        if (clips.length > 0) {
-          const sorted = [...clips].sort((a, b) => a.startMs - b.startMs)
-          selectedClipId = sorted[sorted.length - 1]?.id ?? null
-        }
-      }
-      set({
-        past: [],
-        present,
-        future: [],
-        _lastGroup: null,
-        _lastTs: 0,
-        currentDraft: draft,
-        // UI state — fall back to defaults when the saved draft predates the
-        // wrapped payload shape.
-        presetTab: ui?.presetTab ?? "single",
-        activeLayoutPresetId: ui?.activeLayoutPresetId ?? null,
-        activeCustomPresetId: ui?.activeCustomPresetId ?? null,
-        activeSinglePresetId: ui?.activeSinglePresetId ?? null,
-        bulkEditMode: ui?.bulkEditMode ?? defaultBulk,
-        bulkViewportZoom: ui?.bulkViewportZoom ?? 1,
-        bulkScale: ui?.bulkScale ?? 65,
-        previewAutoScrollDelay: ui?.previewAutoScrollDelay ?? 3000,
-        previewAnimation: ui?.previewAnimation ?? "slide",
-        isAnimateMode: restoreAnimate,
-        selectedAnimationClipId: selectedClipId,
-        selectedAnimationClipIds: selectedClipId ? [selectedClipId] : [],
-        ...CLEAR_SELECTION,
-      })
-    },
-    loadTemplateState: (state, ui) => {
-      const incoming = normalizeEditorState(state)
-      // A template ships a screenshot only so it can render a thumbnail — the
-      // composition (background, frame, shadow, layout…) is what we apply. Drop
-      // every media field and carry over whatever the user already had on their
-      // active canvas, so applying a template restyles their screenshot (or
-      // leaves the canvas empty when they have none).
-      const prev = get().present
-      const prevActive = prev.canvases.find((c) => c.id === prev.activeCanvasId)
-      const present: EditorState = {
-        ...incoming,
-        canvases: incoming.canvases.map((canvas) => {
-          const cleared: CanvasState = {
-            ...canvas,
-            screenshot: null,
-            originalScreenshot: null,
-            lastCropRegion: null,
-            videoClips: null,
-            tweet: null,
-            fullPageCapture: { scrollPosition: 0 },
-            screenshotSlots: canvas.screenshotSlots.map((slot) => ({
-              ...slot,
-              src: null,
-            })),
-          }
-          if (canvas.id === incoming.activeCanvasId && prevActive) {
-            return {
-              ...cleared,
-              screenshot: prevActive.screenshot,
-              originalScreenshot: prevActive.originalScreenshot,
-              lastCropRegion: prevActive.lastCropRegion,
-              videoClips: prevActive.videoClips ?? null,
-              fullPageCapture: prevActive.fullPageCapture ?? {
-                scrollPosition: 0,
-              },
-            }
-          }
-          return cleared
-        }),
-      }
-      const defaultBulk = present.canvases.length > 1
-      const restoreAnimate = Boolean(ui?.isAnimateMode)
-      let selectedClipId: string | null = null
-      if (restoreAnimate) {
-        const active = present.canvases.find(
-          (c) => c.id === present.activeCanvasId
-        )
-        const clips = active?.animation?.clips ?? []
-        if (clips.length > 0) {
-          const sorted = [...clips].sort((a, b) => a.startMs - b.startMs)
-          selectedClipId = sorted[sorted.length - 1]?.id ?? null
-        }
-      }
-      set({
-        past: [],
-        present,
-        future: [],
-        _lastGroup: null,
-        _lastTs: 0,
-        // Templates are a starting point, not a saved project: keep the draft
-        // pointer empty so the first Save writes a new draft.
-        currentDraft: null,
-        presetTab: ui?.presetTab ?? "single",
-        activeLayoutPresetId: ui?.activeLayoutPresetId ?? null,
-        activeCustomPresetId: ui?.activeCustomPresetId ?? null,
-        activeSinglePresetId: ui?.activeSinglePresetId ?? null,
-        bulkEditMode: ui?.bulkEditMode ?? defaultBulk,
-        bulkViewportZoom: ui?.bulkViewportZoom ?? 1,
-        bulkScale: ui?.bulkScale ?? 65,
-        previewAutoScrollDelay: ui?.previewAutoScrollDelay ?? 3000,
-        previewAnimation: ui?.previewAnimation ?? "slide",
-        isAnimateMode: restoreAnimate,
-        selectedAnimationClipId: selectedClipId,
-        selectedAnimationClipIds: selectedClipId ? [selectedClipId] : [],
-        ...CLEAR_SELECTION,
-      })
-    },
-    applyPresetSnapshot: (snapshot, canvasId) => {
-      commit((state) => {
-        const targetId = canvasId ?? state.activeCanvasId
-        const canvases = state.canvases.map((canvas) => {
-          if (canvas.id !== targetId) return canvas
-
-          const existingSlots = canvas.screenshotSlots
-          const slots: ScreenshotSlot[] = canvas.tweet
-            ? []
-            : snapshot.slots.map((config, index) => {
-                const previous = existingSlots[index]
-                // Preset geometry drives pose; live media + per-slot overrides
-                // that presets don't capture (frame, border, padding, …) stay
-                // on the matching index so mixed-frame layouts survive apply.
-                return {
-                  id: previous?.id ?? makeId(),
-                  src: previous?.src ?? null,
-                  originalSrc: previous?.originalSrc ?? null,
-                  lastCropRegion: previous?.lastCropRegion ?? null,
-                  fullPageCapture: previous?.fullPageCapture ?? null,
-                  xPct: config.xPct,
-                  yPct: config.yPct,
-                  widthPct: config.widthPct ?? previous?.widthPct ?? 60,
-                  heightPct: config.heightPct ?? previous?.heightPct ?? 28,
-                  rotation: config.rotation,
-                  tilt: config.tilt,
-                  scale: config.scale,
-                  zIndex:
-                    config.zIndex ??
-                    previous?.zIndex ??
-                    computeNextLayerZ(canvas) + index,
-                  filter: config.filter ?? previous?.filter ?? "none",
-                  adjustments: config.adjustments ?? previous?.adjustments,
-                  hidden: config.hidden ?? previous?.hidden,
-                  objectFit: config.objectFit ?? previous?.objectFit,
-                  shadow: config.shadow ?? previous?.shadow,
-                  border: previous?.border,
-                  borderRadius: previous?.borderRadius,
-                  padding: previous?.padding,
-                  lighting: previous?.lighting,
-                  frame: previous?.frame,
-                }
-              })
-
-          const offset = resolveMainOffsetPx(snapshot.mainOffset)
-
-          const style = snapshot.canvasStyle
-          let next: CanvasState = {
-            // Style bag (background, shadow, layers, …) layered over the live
-            // canvas. mergeCanvasStyle honours the frame-vs-tweet and
-            // tweetSettings special cases and skips absent fields.
-            ...mergeCanvasStyle(canvas, style),
-            // geometry — always driven by the preset, never the style bag
-            tilt: snapshot.canvasTilt,
-            scale: snapshot.canvasScale,
-            screenshotPosition: style?.screenshotPosition ?? "center",
-            screenshotOffset: offset,
-            screenshotSlots: canvas.tweet ? [] : slots,
-            // live pixels are preserved by mergeCanvasStyle (not in the style
-            // bag); kept explicit here for clarity at the apply boundary.
-            screenshot: canvas.screenshot,
-            originalScreenshot: canvas.originalScreenshot,
-            lastCropRegion: canvas.lastCropRegion,
-          }
-
-          // Animate presets replace the timeline and fold the last clip pose
-          // into the committed canvas (matches exit-animate end-frame behavior).
-          if (
-            snapshot.animation &&
-            Array.isArray(snapshot.animation.clips) &&
-            snapshot.animation.clips.length > 0
-          ) {
-            const liveSlotIds = next.screenshotSlots.map((s) => s.id)
-            const remapped = remapAnimationForApply(
-              snapshot.animation,
-              liveSlotIds,
-              next.background
-            )
-            const sorted = [...remapped.clips].sort(
-              (a, b) => a.startMs - b.startMs
-            )
-            const last = sorted[sorted.length - 1]
-            // Older animate presets could save the open final clip before its
-            // live inspector edits had been copied into `clip.pose`. When that
-            // clip owns Position, its stale pose equals its baseline while the
-            // top-level preset geometry still contains the real final placement.
-            // Recover that placement so existing presets animate correctly too.
-            const lastPose = last ? clipPose(last) : null
-            const lastBaseline = last ? clipBaseline(last) : null
-            let repairedAnimation = remapped
-            if (
-              last &&
-              lastPose &&
-              lastBaseline &&
-              (last.effects ?? []).includes("position") &&
-              lastPose.screenshotPosition === lastBaseline.screenshotPosition &&
-              lastPose.screenshotOffset.x === lastBaseline.screenshotOffset.x &&
-              lastPose.screenshotOffset.y === lastBaseline.screenshotOffset.y &&
-              (lastPose.screenshotPosition !== next.screenshotPosition ||
-                lastPose.screenshotOffset.x !== next.screenshotOffset.x ||
-                lastPose.screenshotOffset.y !== next.screenshotOffset.y)
-            ) {
-              repairedAnimation = {
-                ...remapped,
-                clips: remapped.clips.map((clip) =>
-                  clip.id === last.id
-                    ? {
-                        ...clip,
-                        pose: {
-                          ...lastPose,
-                          screenshotPosition: next.screenshotPosition,
-                          screenshotOffset: next.screenshotOffset,
-                        },
-                      }
-                    : clip
-                ),
-              }
-            }
-            // Rest at the animation's final frame, but position sits at its START
-            // (see buildRestingPose) so a "move" preset shows where it begins.
-            const restingPose = buildRestingPose(next, repairedAnimation.clips)
-            const posePatch = restingPose
-              ? applyPoseToCanvas(next, restingPose)
-              : {}
-            next = {
-              ...next,
-              ...posePatch,
-              // Pose may rewrite slots; keep ids from geometry apply.
-              screenshotSlots: (
-                posePatch.screenshotSlots ?? next.screenshotSlots
-              ).map((s, i) => ({
-                ...s,
-                id: next.screenshotSlots[i]?.id ?? s.id,
-                src: next.screenshotSlots[i]?.src ?? s.src,
-              })),
-              animation: repairedAnimation,
-            }
-          }
-
-          return next
-        })
-        return { canvases }
-      }, "preset:apply")
-    },
-    setScreenshot: (screenshot, canvasId) => {
-      commitCanvas(
-        canvasId,
-        (canvas) => ({
-          screenshot,
-          originalScreenshot: screenshot,
-          lastCropRegion: null,
-          fullPageCapture: null,
-          videoClips: null,
-          // A screenshot replaces any tweet as the canvas's main content.
-          tweet: screenshot ? null : canvas.tweet,
-          objectFit: canvas.objectFit ?? "contain",
-          screenshotLayer: {
-            ...canvas.screenshotLayer,
-            zIndex:
-              screenshot && !canvas.screenshot
-                ? computeNextLayerZ(canvas)
-                : canvas.screenshotLayer.zIndex,
-            hidden: false,
-          },
-        }),
-        null
-      )
-    },
-    setFullPageScreenshot: (src, canvasId) => {
-      commitCanvas(
-        canvasId,
-        (canvas) => ({
-          screenshot: src,
-          originalScreenshot: src,
-          lastCropRegion: null,
-          fullPageCapture: src ? { scrollPosition: 0 } : null,
-          videoClips: null,
-          // A URL capture replaces any tweet as the canvas's main content.
-          tweet: src ? null : canvas.tweet,
-          objectFit: canvas.objectFit ?? "contain",
-          screenshotLayer: {
-            ...canvas.screenshotLayer,
-            zIndex:
-              src && !canvas.screenshot
-                ? computeNextLayerZ(canvas)
-                : canvas.screenshotLayer.zIndex,
-            hidden: false,
-          },
-        }),
-        null
-      )
-    },
-    setFullPageScreenshotScrollPosition: (scrollPosition, canvasId) =>
-      commitCanvas(
-        canvasId,
-        (canvas) => ({
-          fullPageCapture: canvas.fullPageCapture
-            ? { scrollPosition: Math.max(0, Math.min(100, scrollPosition)) }
-            : canvas.fullPageCapture,
-        }),
-        "full-page-scroll"
-      ),
-    applyCroppedScreenshot: (s, region, canvasId) =>
-      commitCanvas(
-        canvasId,
-        { screenshot: s, lastCropRegion: region, fullPageCapture: null },
-        "applyCroppedScreenshot"
-      ),
-    setScreenshotCropRegion: (region, canvasId) =>
-      commitCanvasEffect(
-        canvasId,
-        (canvas) => ({
-          lastCropRegion: region,
-          fullPageCapture: region ? null : canvas.fullPageCapture,
-        }),
-        "setScreenshotCropRegion",
-        "crop"
-      ),
-    updateVideoClip: (id, patch, canvasId) =>
-      commitCanvas(
-        canvasId,
-        (canvas) => {
-          const clips = canvas.videoClips ?? [
-            { id: "video-main", timelineStartMs: 0, startMs: 0, endMs: null },
-          ]
-          return {
-            videoClips: clips.map((clip) =>
-              clip.id === id ? { ...clip, ...patch } : clip
-            ),
-          }
-        },
-        "video-trim"
-      ),
-    splitVideoClip: (id, atMs, canvasId) => {
-      let newId: string | null = null
-      commitCanvas(
-        canvasId,
-        (canvas) => {
-          const clips = canvas.videoClips ?? [
-            { id: "video-main", timelineStartMs: 0, startMs: 0, endMs: null },
-          ]
-          const clip = clips.find((item) => item.id === id)
-          if (
-            !clip ||
-            atMs <= clip.startMs ||
-            (clip.endMs !== null && atMs >= clip.endMs)
-          ) {
-            return {}
-          }
-          newId = makeId()
-          return {
-            videoClips: clips.flatMap((item) =>
-              item.id === id
-                ? [
-                    { ...item, endMs: atMs },
-                    {
-                      ...item,
-                      id: newId!,
-                      timelineStartMs:
-                        (item.timelineStartMs ?? item.startMs) +
-                        (atMs - item.startMs),
-                      startMs: atMs,
-                    },
-                  ]
-                : [item]
-            ),
-          }
-        },
-        "video-split"
-      )
-      return newId
-    },
-    duplicateVideoClip: (id, durationMs, canvasId) => {
-      let newId: string | null = null
-      commitCanvas(
-        canvasId,
-        (canvas) => {
-          const clips = canvas.videoClips ?? [
-            { id: "video-main", timelineStartMs: 0, startMs: 0, endMs: null },
-          ]
-          const source = clips.find((clip) => clip.id === id)
-          if (!source || durationMs <= 0) return {}
-          const sourceStart = source.timelineStartMs ?? source.startMs
-          const sourceEndMs = source.endMs ?? source.startMs + durationMs
-          const { startMs, shiftAfterMs, shiftMs } = resolveRippleDrop(
-            sourceStart + durationMs,
-            durationMs,
-            clips
-              .filter((clip) => clip.id !== id)
-              .map((clip) => ({
-                startMs: clip.timelineStartMs ?? clip.startMs,
-                durationMs:
-                  clip.endMs === null ? durationMs : clip.endMs - clip.startMs,
-              })),
-            MAX_DURATION_MS
-          )
-          newId = makeId()
-          return {
-            videoClips: [
-              ...clips.map((clip) => {
-                const timelineStartMs = clip.timelineStartMs ?? clip.startMs
-                const positioned =
-                  timelineStartMs < shiftAfterMs
-                    ? clip
-                    : { ...clip, timelineStartMs: timelineStartMs + shiftMs }
-                return clip.id === id && clip.endMs === null
-                  ? { ...positioned, endMs: sourceEndMs }
-                  : positioned
-              }),
-              {
-                ...source,
-                id: newId,
-                endMs: sourceEndMs,
-                timelineStartMs: startMs,
-              },
-            ],
-          }
-        },
-        "video-duplicate"
-      )
-      return newId
-    },
-    removeVideoClips: (ids, canvasId) =>
-      commitCanvas(
-        canvasId,
-        (canvas) => {
-          const clips = canvas.videoClips ?? [
-            { id: "video-main", timelineStartMs: 0, startMs: 0, endMs: null },
-          ]
-          const kept = clips.filter((clip) => !ids.includes(clip.id))
-          return kept.length > 0
-            ? { videoClips: kept }
-            : {
-                screenshot: null,
-                originalScreenshot: null,
-                videoClips: null,
-                fullPageCapture: null,
-              }
-        },
-        "video-delete"
-      ),
-    setAspect: (a) => {
-      const snapshot = get()
-      commit((state) => {
-        const currentAspect = stateCanvasAspect(state)
-        const nextAspect = aspectRatioFromState(a)
-
-        return {
-          aspect: a,
-          canvases: state.canvases.map((canvas) => {
-            // Resolve the active preset (built-in *or* user-saved custom)
-            // for *this* canvas's frame, so portrait-device variants kick
-            // in correctly for layout presets.
-            const activeGeometry = resolveActivePresetGeometry({
-              activeLayoutPresetId: snapshot.activeLayoutPresetId,
-              activeCustomPresetId: snapshot.activeCustomPresetId,
-              layoutPresets: LAYOUT_PRESETS,
-              customPresets: snapshot.customPresets,
-              frame: canvas.frame,
-            })
-            const shouldReapply =
-              activeGeometry !== null &&
-              canvas.id === state.activeCanvasId &&
-              canvas.screenshotSlots.length === activeGeometry.slots.length
-            const activeSinglePreset =
-              !activeGeometry && canvas.id === state.activeCanvasId
-                ? PRESENT_PRESETS.find(
-                    (preset) => preset.id === snapshot.activeSinglePresetId
-                  )
-                : undefined
-
-            let screenshotSlots = layoutSlotsInRow(
-              canvas.screenshotSlots,
-              canvas.frame,
-              nextAspect
-            )
-            if (shouldReapply && activeGeometry) {
-              screenshotSlots = screenshotSlots.map((naturalSlot, index) => {
-                const config = activeGeometry.slots[index]
-                if (!config) return naturalSlot
-                const { xPct, yPct } = resolveSlotPositionPct({
-                  config,
-                  naturalSlotXPct: naturalSlot.xPct,
-                  relativeSlotPositions: activeGeometry.relativeSlotPositions,
-                })
-                return {
-                  ...naturalSlot,
-                  xPct,
-                  yPct,
-                  rotation: config.rotation,
-                  tilt: config.tilt,
-                  scale: config.scale,
-                  ...(config.zIndex !== undefined && { zIndex: config.zIndex }),
-                }
-              })
-            } else if (activeSinglePreset) {
-              // Single presets track the canvas' current zoom (tilt-only).
-              screenshotSlots = screenshotSlots.map((slot) => ({
-                ...slot,
-                yPct: 50,
-                rotation: 0,
-                tilt: activeSinglePreset.tilt,
-                scale: canvas.scale,
-              }))
-            }
-
-            const screenshotOffset =
-              shouldReapply && activeGeometry
-                ? resolveMainOffsetPx(activeGeometry.mainOffset)
-                : scaleScreenshotOffsetForAspectChange(
-                    canvas.screenshotOffset,
-                    currentAspect,
-                    nextAspect
-                  )
-
-            return {
-              ...canvas,
-              tilt:
-                shouldReapply && activeGeometry
-                  ? activeGeometry.canvasTilt
-                  : activeSinglePreset
-                    ? activeSinglePreset.tilt
-                    : canvas.tilt,
-              scale:
-                shouldReapply && activeGeometry
-                  ? activeGeometry.canvasScale
-                  : activeSinglePreset
-                    ? canvas.scale
-                    : canvas.scale,
-              screenshotOffset,
-              screenshotSlots,
-              annotations: scaleAnnotationStrokesForAspectChange(
-                canvas.annotations,
-                currentAspect,
-                nextAspect
-              ),
-            }
-          }),
-        }
-      }, "aspect")
-    },
-    setCanvasAspect: (canvasId, a) => {
-      const snapshot = get()
-      commitCanvas(
-        canvasId,
-        (canvas, state) => {
-          const currentAspect = aspectRatioFromState(
-            canvas.aspect ?? state.aspect
-          )
-          const nextAspect = aspectRatioFromState(a)
-          const activeGeometry = resolveActivePresetGeometry({
-            activeLayoutPresetId: snapshot.activeLayoutPresetId,
-            activeCustomPresetId: snapshot.activeCustomPresetId,
-            layoutPresets: LAYOUT_PRESETS,
-            customPresets: snapshot.customPresets,
-            frame: canvas.frame,
-          })
-          const shouldReapply =
-            activeGeometry !== null &&
-            canvas.screenshotSlots.length === activeGeometry.slots.length
-          const activeSinglePreset = !activeGeometry
-            ? PRESENT_PRESETS.find(
-                (preset) => preset.id === snapshot.activeSinglePresetId
-              )
-            : undefined
-
-          let screenshotSlots = layoutSlotsInRow(
-            canvas.screenshotSlots,
-            canvas.frame,
-            nextAspect
-          )
-          if (shouldReapply && activeGeometry) {
-            screenshotSlots = screenshotSlots.map((naturalSlot, index) => {
-              const config = activeGeometry.slots[index]
-              if (!config) return naturalSlot
-              const { xPct, yPct } = resolveSlotPositionPct({
-                config,
-                naturalSlotXPct: naturalSlot.xPct,
-                relativeSlotPositions: activeGeometry.relativeSlotPositions,
-              })
-              return {
-                ...naturalSlot,
-                xPct,
-                yPct,
-                rotation: config.rotation,
-                tilt: config.tilt,
-                scale: config.scale,
-                ...(config.zIndex !== undefined && { zIndex: config.zIndex }),
-              }
-            })
-          } else if (activeSinglePreset) {
-            // Single presets track the canvas' current zoom (tilt-only).
-            screenshotSlots = screenshotSlots.map((slot) => ({
-              ...slot,
-              yPct: 50,
-              rotation: 0,
-              tilt: activeSinglePreset.tilt,
-              scale: canvas.scale,
-            }))
-          }
-
-          return {
-            aspect: a,
-            tilt:
-              shouldReapply && activeGeometry
-                ? activeGeometry.canvasTilt
-                : activeSinglePreset
-                  ? activeSinglePreset.tilt
-                  : canvas.tilt,
-            scale:
-              shouldReapply && activeGeometry
-                ? activeGeometry.canvasScale
-                : activeSinglePreset
-                  ? canvas.scale
-                  : canvas.scale,
-            screenshotOffset:
-              shouldReapply && activeGeometry
-                ? resolveMainOffsetPx(activeGeometry.mainOffset)
-                : scaleScreenshotOffsetForAspectChange(
-                    canvas.screenshotOffset,
-                    currentAspect,
-                    nextAspect
-                  ),
-            screenshotSlots,
-            annotations: scaleAnnotationStrokesForAspectChange(
-              canvas.annotations,
-              currentAspect,
-              nextAspect
-            ),
-          }
-        },
-        "aspect"
-      )
-    },
-    setBackground: (b, canvasId, opts) => {
-      // The canvas-view hydration effect re-applies image backgrounds as an
-      // optimized value (library/Unsplash URL → downscaled data URL) on reload.
-      // That's an internal swap of the SAME background, not a user edit, so in
-      // Animate mode it must NOT be recorded as a "background" keyframe effect —
-      // otherwise reopening a draft spuriously marks the open clip as animating
-      // the background. Swap the value on the committed canvas and on any clip
-      // pose holding the same source, leaving `effects` untouched.
-      if (opts?.silent) {
-        commitCanvas(
-          canvasId,
-          (canvas) => {
-            const anim = getCanvasAnimation(canvas)
-            if (anim.clips.length === 0) return { background: b }
-            const prev = canvas.background
-            const matchesPrev = (bg: Background | undefined) =>
-              bg?.type === "image" &&
-              prev.type === "image" &&
-              bg.sourceUrl === prev.sourceUrl
-            const clips = anim.clips.map((c) => {
-              const pose =
-                c.pose && matchesPrev(c.pose.background)
-                  ? { ...c.pose, background: b }
-                  : c.pose
-              const baseline =
-                c.baseline && matchesPrev(c.baseline.background)
-                  ? { ...c.baseline, background: b }
-                  : c.baseline
-              if (pose === c.pose && baseline === c.baseline) return c
-              return { ...c, pose, baseline }
-            })
-            return { background: b, animation: { ...anim, clips } }
-          },
-          "background"
-        )
-        return
-      }
-      commitCanvasEffect(
-        canvasId,
-        { background: b },
-        "background",
-        "background"
-      )
-    },
-    setPadding: (n, canvasId) =>
-      commitCanvasEffect(
-        canvasId,
-        (canvas) => applyScreenshotStyle(canvas, "all", { padding: n }),
-        "padding",
-        "padding"
-      ),
-    setBorderRadius: (n, canvasId) =>
-      commitCanvasEffect(
-        canvasId,
-        (canvas) => applyScreenshotStyle(canvas, "all", { borderRadius: n }),
-        "borderRadius",
-        "borderRadius"
-      ),
-    setCanvasBorderRadius: (n, canvasId) =>
-      commitCanvasEffect(
-        canvasId,
-        { canvasBorderRadius: n },
-        "canvasBorderRadius",
-        "canvasRadius"
-      ),
-    setBorder: (b, canvasId) =>
-      commitCanvasEffect(
-        canvasId,
-        (canvas) => applyScreenshotStyle(canvas, "all", { border: b }),
-        "border",
-        "border"
-      ),
-    setMainScreenshotPadding: (n, canvasId) =>
-      commitCanvasEffect(
-        canvasId,
-        (canvas) => applyScreenshotStyle(canvas, "main", { padding: n }),
-        "padding",
-        "padding"
-      ),
-    setMainScreenshotBorderRadius: (n, canvasId) =>
-      commitCanvasEffect(
-        canvasId,
-        (canvas) => applyScreenshotStyle(canvas, "main", { borderRadius: n }),
-        "borderRadius",
-        "borderRadius"
-      ),
-    setMainScreenshotBorder: (b, canvasId) =>
-      commitCanvasEffect(
-        canvasId,
-        (canvas) => applyScreenshotStyle(canvas, "main", { border: b }),
-        "border",
-        "border"
-      ),
-    setBackdropEffects: (e, canvasId) =>
-      commitCanvasEffect(
-        canvasId,
-        (canvas) => ({ backdrop: { ...canvas.backdrop, effects: e } }),
-        "backdrop-effects",
-        "backdrop"
-      ),
-    setBackdropPattern: (p, canvasId) =>
-      commitCanvasEffect(
-        canvasId,
-        (canvas) => ({ backdrop: { ...canvas.backdrop, pattern: p } }),
-        "backdrop-pattern",
-        "pattern"
-      ),
-    setBackdropLighting: (l, canvasId) =>
-      commitCanvasEffect(
-        canvasId,
-        (canvas) => applyScreenshotStyle(canvas, "all", { lighting: l }),
-        "backdrop-lighting",
-        "lighting"
-      ),
-    setMainScreenshotBackdropLighting: (l, canvasId) =>
-      commitCanvasEffect(
-        canvasId,
-        (canvas) => applyScreenshotStyle(canvas, "main", { lighting: l }),
-        "backdrop-lighting",
-        "lighting"
-      ),
-    setBackdropFilter: (f, canvasId) =>
-      commitCanvasEffect(
-        canvasId,
-        (canvas) => ({ backdrop: { ...canvas.backdrop, filter: f } }),
-        "backdrop-filter",
-        "filter"
-      ),
-    setTilt: (t, canvasId) =>
-      commitCanvasEffect(
-        canvasId,
-        (canvas) => applyScreenshotStyle(canvas, "main", { tilt: t }),
-        "tilt",
-        "tilt"
-      ),
-    setScale: (n, canvasId) =>
-      commitCanvasEffect(
-        canvasId,
-        (canvas) => applyScreenshotStyle(canvas, "main", { scale: n }),
-        "scale",
-        "zoom"
-      ),
-    setTiltAndScale: (t, scale, canvasId) =>
-      commitCanvasEffect(canvasId, { tilt: t, scale }, "tilt-scale", [
-        "tilt",
-        "zoom",
-      ]),
-    setScreenshotTilt: (t, canvasId) =>
-      commitCanvasEffect(
-        canvasId,
-        (canvas) => applyScreenshotStyle(canvas, "all", { tilt: t }),
-        "tilt",
-        "tilt"
-      ),
-    setScreenshotScale: (n, canvasId) =>
-      commitCanvasEffect(
-        canvasId,
-        (canvas) => applyScreenshotStyle(canvas, "all", { scale: n }),
-        "scale",
-        "zoom"
-      ),
-    setScreenshotRotation: (n, canvasId) =>
-      commitCanvasEffect(
-        canvasId,
-        (canvas) => applyScreenshotStyle(canvas, "all", { rotation: n }),
-        "tilt",
-        "tilt"
-      ),
-    setCanvasZoom: (n) => commit({ canvasZoom: n }, "canvasZoom"),
-    setScreenshotPosition: (p, canvasId) =>
-      commitCanvasEffect(
-        canvasId,
-        { screenshotPosition: p, screenshotOffset: { x: 0, y: 0 } },
-        "screenshotPosition",
-        "position"
-      ),
-    setScreenshotOffset: (o, canvasId) =>
-      commitCanvasEffect(
-        canvasId,
-        { screenshotOffset: o },
-        "screenshotOffset",
-        "position"
-      ),
-    setScreenshotPlacement: (p, o, canvasId) =>
-      commitCanvasEffect(
-        canvasId,
-        { screenshotPosition: p, screenshotOffset: o },
-        "screenshotPlacement",
-        "position"
-      ),
-    updateScreenshotLayer: (patch, canvasId) =>
-      commitCanvas(
-        canvasId,
-        (canvas) => ({
-          screenshotLayer: { ...canvas.screenshotLayer, ...patch },
-        }),
-        "screenshotLayer"
-      ),
-    applyScreenshotStyle: (target, patch, canvasId) => {
-      const scope =
-        target === "main" || target === "all" ? target : `slot-${target.slotId}`
-      commitCanvasEffect(
-        canvasId,
-        (canvas) => applyScreenshotStyle(canvas, target, patch),
-        `${screenshotStyleGroup(patch)}:${scope}`,
-        screenshotStyleEffects(patch)
-      )
-    },
-    setShadow: (s, canvasId) =>
-      commitCanvasEffect(
-        canvasId,
-        (canvas) => applyScreenshotStyle(canvas, "all", { shadow: s }),
-        "shadow",
-        "shadow"
-      ),
-    setMainScreenshotShadow: (s, canvasId) =>
-      commitCanvasEffect(
-        canvasId,
-        (canvas) => applyScreenshotStyle(canvas, "main", { shadow: s }),
-        "shadow",
-        "shadow"
-      ),
-    setOverlay: (o, canvasId) =>
-      commitCanvasEffect(canvasId, { overlay: o }, "overlay", "overlay"),
-    setFrame: (f, canvasId) =>
-      commitCanvas(
-        canvasId,
-        (canvas, state) => {
-          if (canvas.tweet) {
-            return {
-              frame: { id: "none", color: "black", orientation: "vertical" },
-              frameAddress: "",
-            }
-          }
-          return applySharedFrameToCanvas(
-            canvas,
-            state,
-            f,
-            get().activeLayoutPresetId
-          )
-        },
-        "frame"
-      ),
-    setFrameForMatchingScreenshots: (f, canvasId) =>
-      commitCanvas(
-        canvasId,
-        (canvas, state) => {
-          if (canvas.tweet) {
-            return {
-              frame: { id: "none", color: "black", orientation: "vertical" },
-              frameAddress: "",
-            }
-          }
-          // Strip overrides *before* reflow so row math packs everyone as the
-          // shared frame (not the pre-apply mixed-frame widths).
-          const withoutSlotFrames: CanvasState = {
-            ...canvas,
-            screenshotSlots: canvas.screenshotSlots.map((slot) =>
-              slot.frame ? { ...slot, frame: undefined } : slot
-            ),
-          }
-          return applySharedFrameToCanvas(
-            withoutSlotFrames,
-            state,
-            f,
-            get().activeLayoutPresetId
-          )
-        },
-        "frame"
-      ),
-    setMainScreenshotFrame: (f, canvasId) =>
-      commitCanvas(
-        canvasId,
-        (canvas) => {
-          if (canvas.tweet) {
-            return {
-              frame: { id: "none", color: "black", orientation: "vertical" },
-              frameAddress: "",
-            }
-          }
-          // The main has no frame of its own — it uses the canvas frame, which
-          // every un-overridden slot inherits. To change ONLY the main, pin each
-          // inheriting slot to the previous frame so it stays put.
-          const previousFrame = canvas.frame
-          return {
-            frame: { ...f },
-            screenshotSlots: canvas.screenshotSlots.map((slot) =>
-              slot.frame ? slot : { ...slot, frame: { ...previousFrame } }
-            ),
-          }
-        },
-        "frame"
-      ),
-    setFrameAddress: (address, canvasId) =>
-      commitCanvas(canvasId, { frameAddress: address }, "frame-address"),
-    setTweet: (card, canvasId) => {
-      commit((state) => {
-        const targetId = canvasId ?? state.activeCanvasId
-        return {
-          aspect: { ...TWEET_POST_ASPECT },
-          canvases: state.canvases.map((canvas) =>
-            canvas.id === targetId
-              ? {
-                  ...canvas,
-                  // A tweet replaces the screenshot as the canvas's main content.
-                  tweet: card,
-                  screenshot: null,
-                  originalScreenshot: null,
-                  lastCropRegion: null,
-                  fullPageCapture: null,
-                  videoClips: null,
-                  screenshotSlots: [],
-                  frame: {
-                    id: "none",
-                    color: "black",
-                    orientation: "vertical",
-                  },
-                  frameAddress: "",
-                  screenshotPosition: "center",
-                  screenshotOffset: { x: 0, y: 0 },
-                  aspect: undefined,
-                }
-              : canvas
-          ),
-        }
-      }, null)
-      set({
-        presetTab: "single",
-        activeLayoutPresetId: null,
-        activeCustomPresetId: null,
-      })
-    },
-    updateTweet: (patch, canvasId) =>
-      commitCanvas(
-        canvasId,
-        (canvas) =>
-          canvas.tweet ? { tweet: { ...canvas.tweet, ...patch } } : {},
-        "tweet"
-      ),
-    clearTweet: (canvasId) => commitCanvas(canvasId, { tweet: null }, null),
-    setObjectFit: (fit, canvasId) =>
-      commitCanvas(canvasId, { objectFit: fit }, "objectFit"),
-    bringScreenshotToFront: (canvasId) =>
-      commitCanvas(
-        canvasId,
-        (canvas) => moveLayerInStack(canvas, "screenshot", "front"),
-        "screenshot-layer"
-      ),
-    sendScreenshotToBack: (canvasId) =>
-      commitCanvas(
-        canvasId,
-        (canvas) => moveLayerInStack(canvas, "screenshot", "back"),
-        "screenshot-layer"
-      ),
-    setPortrait: (p, canvasId) =>
-      commitCanvasEffect(canvasId, { portrait: p }, "portrait", "portrait"),
-    setEnhance: (e, canvasId) =>
-      commitCanvas(canvasId, { enhance: e }, "enhance"),
-    setAnnotation: (patch) =>
-      commit(
-        (state) => ({ annotation: { ...state.annotation, ...patch } }),
-        "annotation"
-      ),
-
-    addAnnotationStroke: (stroke, canvasId) => {
-      const id = makeId()
-      commitCanvas(
-        canvasId,
-        (canvas) => ({
-          annotations: [
-            ...canvas.annotations,
-            { ...stroke, id, zIndex: computeNextLayerZ(canvas) },
-          ],
-        }),
-        `annotation-stroke-${id}`
-      )
-      return id
-    },
-    updateAnnotationStroke: (id, points, canvasId) =>
-      commitCanvas(
-        canvasId,
-        (canvas) => ({
-          annotations: canvas.annotations.map((stroke) =>
-            stroke.id === id ? { ...stroke, points } : stroke
-          ),
-        }),
-        `annotation-stroke-${id}`
-      ),
-    updateAnnotationStrokeLayer: (id, patch, canvasId) =>
-      commitCanvas(
-        canvasId,
-        (canvas) => ({
-          annotations: canvas.annotations.map((stroke) =>
-            stroke.id === id ? { ...stroke, ...patch } : stroke
-          ),
-        }),
-        `annotation-stroke-${id}`
-      ),
-    deleteAnnotationStroke: (id, canvasId) =>
-      commitCanvas(
-        canvasId,
-        (canvas) => ({
-          annotations: canvas.annotations.filter((stroke) => stroke.id !== id),
-        }),
-        null
-      ),
-
-    addAnnotationShape: (shape, canvasId) => {
-      const id = makeId()
-      commitCanvas(
-        canvasId,
-        (canvas) => ({
-          annotationShapes: [
-            ...canvas.annotationShapes,
-            { ...shape, id, zIndex: computeNextLayerZ(canvas) },
-          ],
-        }),
-        null
-      )
-      return id
-    },
-    updateAnnotationShape: (id, patch, canvasId) =>
-      commitCanvas(
-        canvasId,
-        (canvas) => ({
-          annotationShapes: canvas.annotationShapes.map((shape) =>
-            shape.id === id ? { ...shape, ...patch } : shape
-          ),
-        }),
-        `annotation-shape-${id}`
-      ),
-    deleteAnnotationShape: (id, canvasId) =>
-      commitCanvas(
-        canvasId,
-        (canvas) => ({
-          annotationShapes: canvas.annotationShapes.filter(
-            (shape) => shape.id !== id
-          ),
-        }),
-        null
-      ),
-    duplicateAnnotationShape: (id, canvasId) => {
-      const copyId = makeId()
-      let didCopy = false
-      commitCanvas(
-        canvasId,
-        (canvas) => {
-          const result = duplicateLayerItem(
-            canvas.annotationShapes,
-            id,
-            copyId,
-            computeNextLayerZ(canvas),
-            { offset: 3, maxPct: 98 }
-          )
-          didCopy = result.ok
-          return { annotationShapes: result.items }
-        },
-        null
-      )
-      return didCopy ? copyId : null
-    },
-    bringAnnotationShapeToFront: annotationShapeLayerOps.toFront,
-    sendAnnotationShapeToBack: annotationShapeLayerOps.toBack,
-    clearAnnotations: (canvasId) =>
-      commitCanvas(canvasId, { annotations: [], annotationShapes: [] }, null),
-
-    addText: (canvasId) => {
-      const id = makeId()
-      const state = get()
-      const aw = state.present.aspect.w || 16
-      const ah = state.present.aspect.h || 10
-      const canvasW = CANVAS_BASE_W
-      const canvasH = (CANVAS_BASE_W * ah) / aw
-      const defaultFontSize = Math.round(
-        Math.min(96, Math.max(18, Math.max(canvasW, canvasH) * 0.028))
-      )
-      commitCanvas(
-        canvasId,
-        (canvas) => ({
-          texts: [
-            ...canvas.texts,
-            {
-              id,
-              content: "Double-click to edit",
-              xPct: 50,
-              yPct: 85,
-              rotation: 0,
-              fontSize: defaultFontSize,
-              fontFamily: FONT_FAMILIES[0].css,
-              fontWeight: 500,
-              lineHeight: 1.3,
-              letterSpacing: 0,
-              color: "#ffffff",
-              align: "left",
-              borderColor: null,
-              borderWidth: 1,
-              borderStyle: "solid",
-              zIndex: computeNextLayerZ(canvas),
-              widthPx: null,
-              heightPx: null,
-              autoColor: true,
-              opacity: 100,
-              blendMode: "normal",
-            },
-          ],
-        }),
-        null
-      )
-      return id
-    },
-    updateText: (id, patch, canvasId) =>
-      commitCanvas(
-        canvasId,
-        (canvas) => ({
-          texts: canvas.texts.map((t) =>
-            t.id === id ? { ...t, ...patch } : t
-          ),
-        }),
-        `text-${id}`
-      ),
-    deleteText: (id, canvasId) =>
-      commitCanvas(
-        canvasId,
-        (canvas) => ({ texts: canvas.texts.filter((t) => t.id !== id) }),
-        null
-      ),
-    duplicateText: (id, canvasId) => {
-      const copyId = makeId()
-      let didCopy = false
-      commitCanvas(
-        canvasId,
-        (canvas) => {
-          const result = duplicateLayerItem(
-            canvas.texts,
-            id,
-            copyId,
-            computeNextLayerZ(canvas)
-          )
-          didCopy = result.ok
-          return { texts: result.items }
-        },
-        null
-      )
-      return didCopy ? copyId : null
-    },
-    bringTextToFront: textLayerOps.toFront,
-    sendTextToBack: textLayerOps.toBack,
-    setSelectedTextId: (id) =>
-      set(
-        id
-          ? { ...CLEAR_SELECTION, selectedTextId: id }
-          : { selectedTextId: null }
-      ),
-
-    addAsset: (src, canvasId) => {
-      const id = makeId()
-      commitCanvas(
-        canvasId,
-        (canvas) => ({
-          assets: [
-            ...canvas.assets,
-            {
-              id,
-              src,
-              xPct: 50,
-              yPct: 50,
-              widthPct: 25,
-              heightPct: null,
-              rotation: 0,
-              zIndex: computeNextLayerZ(canvas),
-              opacity: 100,
-              filter: "none",
-              blendMode: "normal",
-              hidden: false,
-            },
-          ],
-        }),
-        null
-      )
-      return id
-    },
-    updateAsset: (id, patch, canvasId) =>
-      commitCanvas(
-        canvasId,
-        (canvas) => ({
-          assets: canvas.assets.map((a) =>
-            a.id === id ? { ...a, ...patch } : a
-          ),
-        }),
-        `asset-${id}`
-      ),
-    deleteAsset: (id, canvasId) =>
-      commitCanvas(
-        canvasId,
-        (canvas) => ({ assets: canvas.assets.filter((a) => a.id !== id) }),
-        null
-      ),
-    duplicateAsset: (id, canvasId) => {
-      const copyId = makeId()
-      let didCopy = false
-      commitCanvas(
-        canvasId,
-        (canvas) => {
-          const result = duplicateLayerItem(
-            canvas.assets,
-            id,
-            copyId,
-            computeNextLayerZ(canvas)
-          )
-          didCopy = result.ok
-          return { assets: result.items }
-        },
-        null
-      )
-      return didCopy ? copyId : null
-    },
-    bringAssetToFront: assetLayerOps.toFront,
-    sendAssetToBack: assetLayerOps.toBack,
-    setSelectedAssetId: (id) =>
-      set(
-        id
-          ? { ...CLEAR_SELECTION, selectedAssetId: id }
-          : { selectedAssetId: null }
-      ),
-    setSelectedAnnotationShapeId: (id) =>
-      set(
-        id
-          ? { ...CLEAR_SELECTION, selectedAnnotationShapeId: id }
-          : { selectedAnnotationShapeId: null }
-      ),
-    setSelectedScreenshotSlotId: (id) =>
-      set(
-        id
-          ? { ...CLEAR_SELECTION, selectedScreenshotSlotId: id }
-          : { selectedScreenshotSlotId: null }
-      ),
-    setIsScreenshotSelected: (selected) =>
-      set(
-        selected
-          ? { ...CLEAR_SELECTION, isScreenshotSelected: true }
-          : { isScreenshotSelected: false }
-      ),
-    setTopBarPopoverOpen: (open) => set({ topBarPopoverOpen: open }),
-    setIsAnimateMode: (a) => {
-      const state = get()
-      const canvas = state.present.canvases.find(
-        (c) => c.id === state.present.activeCanvasId
-      )
-      if (!canvas) {
-        set({
-          isAnimateMode: a,
-          selectedAnimationClipId: null,
-          selectedAnimationClipIds: [],
-        })
-        return
-      }
-      const animation = getCanvasAnimation(canvas)
-      const sorted = [...animation.clips].sort((x, y) => x.startMs - y.startMs)
-      const last = sorted[sorted.length - 1]
-      if (a) {
-        // Entering: the committed canvas is the final frame for every effect
-        // except position (which rests at its START). Fold any edits made outside
-        // Animate mode into the last clip's pose and open it for editing.
-        if (!last) {
-          set({
-            isAnimateMode: true,
-            selectedAnimationClipId: null,
-            selectedAnimationClipIds: [],
-          })
-          return
-        }
-        const pose = captureClipPose(canvas)
-        // The committed canvas holds the animation's START, so folding it
-        // wholesale would flatten every keyframe into its own origin. Keep the
-        // last clip's stored pose for anything a keyframe animates; only a
-        // property nothing animates carries an outside edit onto that keyframe.
-        const foldedPose = overlayEffectFields(
-          pose,
-          clipPose(last),
-          sorted,
-          (c, e) => clipOwnsEffect(c, e)
-        )
-        // The outside edit belongs at the animation's start instead: each
-        // effect's FIRST owning keyframe is the one that eases from it.
-        let nextClips = animation.clips.map((c) =>
-          c.id === last.id ? { ...c, pose: foldedPose } : c
-        )
-        nextClips = nextClips.map((c) => {
-          const startsHere = ANIMATION_EFFECTS.filter(
-            (e) =>
-              clipOwnsEffect(c, e) &&
-              sorted.find((o) => clipOwnsEffect(o, e))?.id === c.id
-          )
-          if (startsHere.length === 0) return c
-          return {
-            ...c,
-            baseline: overlayEffectFields(clipBaseline(c), pose, [c], (_c, e) =>
-              startsHere.includes(e)
-            ),
-          }
-        })
-        // Load the open (last) keyframe's pose so editing it works on its own
-        // values rather than on the resting start the canvas is showing.
-        const canvasPatch = applyPoseToCanvas(canvas, foldedPose)
-        const canvases = state.present.canvases.map((c) =>
-          c.id === canvas.id
-            ? {
-                ...c,
-                ...canvasPatch,
-                animation: { ...animation, clips: nextClips },
-              }
-            : c
-        )
-        set({
-          present: { ...state.present, canvases },
-          isAnimateMode: true,
-          selectedAnimationClipId: last.id,
-          selectedAnimationClipIds: [last.id],
-        })
-        return
-      }
-      // Exiting: persist the open clip's edits, then restore the committed canvas
-      // to the resting pose — where the animation STARTS (see buildRestingPose).
-      // Keyframes describe motion; they must not leave their look baked into the
-      // document, so the static editor and the still export show the composition
-      // the user built. Clear the clip selection.
-      const openId = state.selectedAnimationClipId
-      let nextClips = animation.clips
-      if (openId && nextClips.some((c) => c.id === openId)) {
-        const pose = captureClipPose(canvas)
-        nextClips = nextClips.map((c) => (c.id === openId ? { ...c, pose } : c))
-      }
-      const restingPose = buildRestingPose(canvas, nextClips)
-      const canvasPatch = restingPose
-        ? applyPoseToCanvas(canvas, restingPose)
-        : {}
-      const canvases = state.present.canvases.map((c) =>
-        c.id === canvas.id
-          ? {
-              ...c,
-              ...canvasPatch,
-              animation: { ...animation, clips: nextClips },
-            }
-          : c
-      )
-      set({
-        present: { ...state.present, canvases },
-        isAnimateMode: false,
-        selectedAnimationClipId: null,
-        selectedAnimationClipIds: [],
-      })
-    },
-    selectAnimationClip: (id, canvasId) => {
-      const state = get()
-      // Re-selecting the already-open clip is a no-op for the canvas: its live
-      // edits are in the committed canvas; reloading its (not-yet-saved) stored
-      // pose would wipe them. onClipPointerDown re-selects on every click, so
-      // this matters. Still collapse any multi-selection down to just this clip.
-      if (id === state.selectedAnimationClipId) {
-        const next = id ? [id] : []
-        const cur = state.selectedAnimationClipIds
-        if (cur.length !== next.length || cur[0] !== next[0]) {
-          set({ selectedAnimationClipIds: next })
-        }
-        return
-      }
-      const targetCanvasId = canvasId ?? state.present.activeCanvasId
-      const canvas = state.present.canvases.find((c) => c.id === targetCanvasId)
-      if (!canvas) {
-        set({
-          selectedAnimationClipId: id,
-          selectedAnimationClipIds: id ? [id] : [],
-        })
-        return
-      }
-      const animation = getCanvasAnimation(canvas)
-      const openId = state.selectedAnimationClipId
-      let nextClips = animation.clips
-      // Persist the previously-open clip's edits from the live canvas.
-      if (openId && openId !== id && nextClips.some((c) => c.id === openId)) {
-        const pose = captureClipPose(canvas)
-        nextClips = nextClips.map((c) => (c.id === openId ? { ...c, pose } : c))
-      }
-      // Load the resolved look AT this keyframe (its owned effects + those held
-      // from earlier keyframes) so the inspector/canvas show what it really looks
-      // like there and you can edit from that state.
-      const opened = id ? nextClips.find((c) => c.id === id) : undefined
-      const canvasPatch = opened
-        ? applyPoseToCanvas(
-            canvas,
-            resolveKeyframePose(canvas, nextClips, opened)
-          )
-        : {}
-      const canvases = state.present.canvases.map((c) =>
-        c.id === canvas.id
-          ? {
-              ...c,
-              ...canvasPatch,
-              animation: { ...animation, clips: nextClips },
-            }
-          : c
-      )
-      // Point the inspector at the screenshot this clip targets, so edits route
-      // to the right screenshot (main vs a slot) and get recorded as this clip's
-      // effects. A slot target that no longer exists falls back to the main.
-      const targetSelection = (() => {
-        const t = opened?.target ?? { scope: "all" as const }
-        if (
-          t.scope === "slot" &&
-          canvas.screenshotSlots.some((s) => s.id === t.slotId)
-        ) {
-          return { ...CLEAR_SELECTION, selectedScreenshotSlotId: t.slotId }
-        }
-        if (t.scope === "main") {
-          return { ...CLEAR_SELECTION, isScreenshotSelected: true }
-        }
-        return { ...CLEAR_SELECTION }
-      })()
-      // Route through raw `set` (not commit) so navigating between clips does not
-      // pile up undo history; property edits still commit normally.
-      set({
-        present: { ...state.present, canvases },
-        selectedAnimationClipId: id,
-        selectedAnimationClipIds: id ? [id] : [],
-        ...(opened ? targetSelection : {}),
-      })
-    },
-    setAnimationDuration: (ms, canvasId) =>
-      commitCanvas(
-        canvasId,
-        (canvas) => ({
-          animation: { ...getCanvasAnimation(canvas), durationMs: ms },
-        }),
-        "animation-duration"
-      ),
-    addAnimationClip: (canvasId, atMs) => {
-      const id = makeId()
-      commitCanvas(
-        canvasId,
-        (canvas) => {
-          const animation = getCanvasAnimation(canvas)
-          const clipLen = Math.min(DEFAULT_CLIP_DURATION_MS, MAX_DURATION_MS)
-          const sorted = [...animation.clips].sort(
-            (a, b) => a.startMs - b.startMs
-          )
-          let startMs: number
-          if (atMs != null) {
-            // Drop the clip where the pointer released — allowed past the set
-            // duration (clamped only to the absolute max range).
-            const maxStart = Math.max(0, MAX_DURATION_MS - clipLen)
-            startMs = Math.max(0, Math.min(maxStart, atMs))
-          } else {
-            // Append right after the last clip — allowed to run PAST the set
-            // duration (clamped only to the absolute max range). Clips beyond the
-            // duration are shown faded in the timeline to signal they won't play
-            // until the duration is extended.
-            const lastEnd = sorted.reduce(
-              (max, clip) => Math.max(max, clip.startMs + clip.durationMs),
-              0
-            )
-            startMs = Math.min(lastEnd, Math.max(0, MAX_DURATION_MS - clipLen))
-          }
-          // Never overlap a neighbouring clip: keep the start past the previous
-          // clip's end, and shrink the duration so it stops at the next clip.
-          const prevEnd = sorted
-            .filter((c) => c.startMs <= startMs)
-            .reduce((max, c) => Math.max(max, c.startMs + c.durationMs), 0)
-          startMs = Math.max(startMs, prevEnd)
-          const nextStart = sorted
-            .filter((c) => c.startMs >= startMs)
-            .reduce((min, c) => Math.min(min, c.startMs), MAX_DURATION_MS)
-          const fittedDuration = Math.max(
-            MIN_ANIMATION_CLIP_MS,
-            Math.min(clipLen, nextStart - startMs)
-          )
-          const snapshot = captureClipPose(canvas)
-          const clip: AnimationClip = {
-            id,
-            startMs,
-            durationMs: fittedDuration,
-            target: resolveSelectionTarget(
-              canvas,
-              get().selectedScreenshotSlotId,
-              get().isScreenshotSelected
-            ),
-            pose: snapshot,
-            // The state BEFORE this keyframe's edits — captured at creation so an
-            // effect (e.g. background) can cross-fade FROM the pre-edit value
-            // rather than from a neutral/black origin. The canvas always has a
-            // background, so the first background swap starts from this one.
-            baseline: snapshot,
-            // A fresh keyframe owns nothing until you edit an effect on it.
-            effects: [],
-            // Explicit linear so undefined easing can keep meaning historic
-            // ease-out for drafts/templates saved before per-clip easing.
-            easing: NEW_CLIP_EASING,
-          }
-          return {
-            animation: { ...animation, clips: [...animation.clips, clip] },
-          }
-        },
-        null
-      )
-      return id
-    },
-    updateAnimationClip: (id, patch, canvasId) =>
-      commitCanvas(
-        canvasId,
-        (canvas) => {
-          const animation = getCanvasAnimation(canvas)
-          return {
-            animation: {
-              ...animation,
-              clips: animation.clips.map((clip) =>
-                clip.id === id ? { ...clip, ...patch } : clip
-              ),
-            },
-          }
-        },
-        `animation-clip:${id}`
-      ),
-    clearAnimationClipEffects: (id, canvasId) =>
-      commitCanvas(
-        canvasId,
-        (canvas) => {
-          const animation = getCanvasAnimation(canvas)
-          const nextClips = clearClipEffectsInArray(animation.clips, id)
-          // Same array back → nothing owned, nothing to strip.
-          if (nextClips === animation.clips) return {}
-          const cleared = nextClips.find((c) => c.id === id)
-          // If this clip is open for editing, the committed canvas is showing its
-          // (now-removed) effects — reload the resolved look WITHOUT this clip so
-          // the canvas reflects the strip (e.g. the lit backdrop goes dark).
-          const isOpen = get().selectedAnimationClipId === id
-          const canvasPatch =
-            isOpen && cleared
-              ? applyPoseToCanvas(
-                  canvas,
-                  resolveKeyframePose(canvas, nextClips, cleared)
-                )
-              : {}
-          return {
-            ...canvasPatch,
-            animation: { ...animation, clips: nextClips },
-          }
-        },
-        `animation-clip:${id}`
-      ),
-    removeAnimationClip: (id, canvasId) =>
-      commitCanvas(
-        canvasId,
-        (canvas) => {
-          const animation = getCanvasAnimation(canvas)
-          // Just drop the clip — the others keep their positions. The removed
-          // clip fades out via AnimatePresence in the timeline view.
-          return {
-            animation: {
-              ...animation,
-              clips: animation.clips.filter((clip) => clip.id !== id),
-            },
-          }
-        },
-        null
-      ),
-    moveAnimationClip: (id, startMs, canvasId) =>
-      commitCanvas(
-        canvasId,
-        (canvas) => {
-          const animation = getCanvasAnimation(canvas)
-          const moving = animation.clips.find((clip) => clip.id === id)
-          if (!moving) return {}
-          const dur = moving.durationMs
-          const others = animation.clips.filter((clip) => clip.id !== id)
-          const {
-            startMs: start,
-            shiftAfterMs,
-            shiftMs: shift,
-          } = resolveRippleDrop(startMs, dur, others, MAX_DURATION_MS)
-          const clips = animation.clips.map((clip) => {
-            if (clip.id === id) return { ...clip, startMs: start }
-            if (clip.startMs < shiftAfterMs) return clip
-            return {
-              ...clip,
-              startMs: Math.min(
-                clip.startMs + shift,
-                Math.max(0, MAX_DURATION_MS - clip.durationMs)
-              ),
-            }
-          })
-          // Duration is user-controlled via the end handle — moving a clip never
-          // grows it; clips past the duration are just shown faded.
-          return { animation: { ...animation, clips } }
-        },
-        // Merge with the drag's history entries so the whole gesture undoes once.
-        `animation-clip:${id}`
-      ),
-    duplicateAnimationClip: (id, canvasId) => {
-      const state = get().present
-      const resolvedId = canvasId ?? state.activeCanvasId
-      const canvas = state.canvases.find((c) => c.id === resolvedId)
-      const source = canvas
-        ? getCanvasAnimation(canvas).clips.find((clip) => clip.id === id)
-        : undefined
-      if (!source) return null
-      const newId = makeId()
-      commitCanvas(
-        canvasId,
-        (c) => {
-          const animation = getCanvasAnimation(c)
-          // The copy sits immediately after the original (clamped to the max
-          // range; it may land past the set duration and render faded). Duration
-          // is user-controlled — the copy never grows it.
-          return {
-            animation: {
-              ...animation,
-              clips: insertClipCopy(animation.clips, source.id, newId),
-            },
-          }
-        },
-        null
-      )
-      return newId
-    },
-    setAnimationClipSelection: (ids, canvasId) => {
-      const unique = Array.from(new Set(ids))
-      // 0 or 1 clip → identical to a normal single select (loads its pose so the
-      // inspector edits that keyframe, or deselects on empty).
-      if (unique.length <= 1) {
-        get().selectAnimationClip(unique[0] ?? null, canvasId)
-        return
-      }
-      const state = get()
-      const targetCanvasId = canvasId ?? state.present.activeCanvasId
-      const canvas = state.present.canvases.find((c) => c.id === targetCanvasId)
-      if (!canvas) {
-        set({ selectedAnimationClipIds: unique, selectedAnimationClipId: null })
-        return
-      }
-      const animation = getCanvasAnimation(canvas)
-      const openId = state.selectedAnimationClipId
-      let nextClips = animation.clips
-      // Persist the currently-open clip's live edits before clearing the primary.
-      if (openId && nextClips.some((c) => c.id === openId)) {
-        const pose = captureClipPose(canvas)
-        nextClips = nextClips.map((c) => (c.id === openId ? { ...c, pose } : c))
-      }
-      const canvases = state.present.canvases.map((c) =>
-        c.id === canvas.id
-          ? { ...c, animation: { ...animation, clips: nextClips } }
-          : c
-      )
-      // A multi-selection opens no single keyframe for editing (primary = null)
-      // so inspector edits can't mis-route; bulk actions read the id set.
-      set({
-        present: { ...state.present, canvases },
-        selectedAnimationClipIds: unique,
-        selectedAnimationClipId: null,
-      })
-    },
-    removeAnimationClips: (ids, canvasId) => {
-      if (ids.length === 0) return
-      const idSet = new Set(ids)
-      commitCanvas(
-        canvasId,
-        (canvas) => {
-          const animation = getCanvasAnimation(canvas)
-          return {
-            animation: {
-              ...animation,
-              clips: animation.clips.filter((clip) => !idSet.has(clip.id)),
-            },
-          }
-        },
-        null
-      )
-      // Drop the removed ids from the selection.
-      set((s) => ({
-        selectedAnimationClipIds: s.selectedAnimationClipIds.filter(
-          (id) => !idSet.has(id)
-        ),
-        selectedAnimationClipId:
-          s.selectedAnimationClipId && idSet.has(s.selectedAnimationClipId)
-            ? null
-            : s.selectedAnimationClipId,
-      }))
-    },
-    clearAnimationClipsEffects: (ids, canvasId) => {
-      if (ids.length === 0) return
-      commitCanvas(
-        canvasId,
-        (canvas) => {
-          const animation = getCanvasAnimation(canvas)
-          let nextClips = animation.clips
-          for (const id of ids)
-            nextClips = clearClipEffectsInArray(nextClips, id)
-          // No clip owned anything → nothing changed.
-          if (nextClips === animation.clips) return {}
-          // If the open clip was among those cleared, reload its resolved look
-          // so the committed canvas reflects the strip.
-          const openId = get().selectedAnimationClipId
-          const opened =
-            openId && ids.includes(openId)
-              ? nextClips.find((c) => c.id === openId)
-              : undefined
-          const canvasPatch = opened
-            ? applyPoseToCanvas(
-                canvas,
-                resolveKeyframePose(canvas, nextClips, opened)
-              )
-            : {}
-          return {
-            ...canvasPatch,
-            animation: { ...animation, clips: nextClips },
-          }
-        },
-        null
-      )
-    },
-    duplicateAnimationClips: (ids, canvasId) => {
-      if (ids.length === 0) return []
-      const state = get().present
-      const resolvedId = canvasId ?? state.activeCanvasId
-      const canvas = state.canvases.find((c) => c.id === resolvedId)
-      if (!canvas) return []
-      const existing = getCanvasAnimation(canvas).clips
-      // Duplicate in timeline order so each copy lands right after its original.
-      const sources = ids
-        .map((id) => existing.find((c) => c.id === id))
-        .filter((c): c is AnimationClip => Boolean(c))
-        .sort((a, b) => a.startMs - b.startMs)
-      if (sources.length === 0) return []
-      const newIds: string[] = []
-      commitCanvas(
-        canvasId,
-        (c) => {
-          const animation = getCanvasAnimation(c)
-          let clips = animation.clips
-          for (const source of sources) {
-            const newId = makeId()
-            newIds.push(newId)
-            clips = insertClipCopy(clips, source.id, newId)
-          }
-          return { animation: { ...animation, clips } }
-        },
-        null
-      )
-      return newIds
-    },
-    splitAnimationClip: (id, atMs, canvasId) => {
-      const state = get().present
-      const resolvedId = canvasId ?? state.activeCanvasId
-      const canvas = state.canvases.find((c) => c.id === resolvedId)
-      const source = canvas
-        ? getCanvasAnimation(canvas).clips.find((clip) => clip.id === id)
-        : undefined
-      if (!source) return null
-      const end = source.startMs + source.durationMs
-      // A cut has to leave a real clip on each side, so the clip must be at
-      // least twice the minimum length. Anything shorter simply can't be split.
-      if (source.durationMs < MIN_ANIMATION_CLIP_MS * 2) {
-        return null
-      }
-      // Clamp the cut into the legal window (≥ MIN from each edge) so a click
-      // near an edge still cuts at the nearest valid point instead of silently
-      // doing nothing.
-      atMs = Math.min(
-        Math.max(atMs, source.startMs + MIN_ANIMATION_CLIP_MS),
-        end - MIN_ANIMATION_CLIP_MS
-      )
-      // When the clip being cut is the one open for editing, its live edits sit
-      // on the committed canvas (not yet folded into its stored pose), so capture
-      // the canvas as the true target keyframe. "Open" status then transfers to
-      // the second half below so the follow-up reselect doesn't overwrite the
-      // first half's cut pose with the live one.
-      const wasOpen = get().selectedAnimationClipId === id
-      const newId = makeId()
-      commitCanvas(
-        canvasId,
-        (c, state) => {
-          const animation = getCanvasAnimation(c)
-          const src = animation.clips.find((clip) => clip.id === id)
-          if (!src) return {}
-          const srcEnd = src.startMs + src.durationMs
-          // The clip's own target keyframe (what it eases TO)...
-          const toPose = wasOpen ? captureClipPose(c) : clipPose(src)
-          // ...and the pose it eases FROM — the accumulated state the instant
-          // before it starts, resolved with the same logic playback uses (ask
-          // for the keyframe pose just before this clip's start).
-          const fromPose = resolveKeyframePose(c, animation.clips, {
-            ...src,
-            startMs: src.startMs - 1,
-          })
-          const affectedSlotIds = c.screenshotSlots
-            .filter((s) => clipAffectsSlot(src, s.id))
-            .map((s) => s.id)
-          // Main position eases in percent-point space (as playback does), then
-          // inverts back to a (cell, offset) keyframe pinned to the target cell.
-          const aspect = c.aspect ?? state.aspect
-          const positionAt = (easedP: number) => {
-            const slots = c.screenshotSlots
-            const fromPt = mainPositionPoint(
-              aspect,
-              c.frame,
-              fromPose.screenshotPosition,
-              fromPose.screenshotOffset,
-              slots
-            )
-            const toPt = mainPositionPoint(
-              aspect,
-              c.frame,
-              toPose.screenshotPosition,
-              toPose.screenshotOffset,
-              slots
-            )
-            const midPt = {
-              xPct: fromPt.xPct + (toPt.xPct - fromPt.xPct) * easedP,
-              yPct: fromPt.yPct + (toPt.yPct - fromPt.yPct) * easedP,
-            }
-            const cell = toPose.screenshotPosition
-            return {
-              screenshotPosition: cell,
-              screenshotOffset: mainPositionOffsetForPoint(
-                aspect,
-                c.frame,
-                cell,
-                slots,
-                midPt
-              ),
-            }
-          }
-          // The eased pose at the cut becomes the boundary keyframe: the first
-          // half eases from → cut, the second half eases cut → to, so each piece
-          // plays its portion of the original motion instead of the whole thing.
-          const midPose = poseAtCut(
-            fromPose,
-            toPose,
-            (atMs - src.startMs) / src.durationMs,
-            src.effects ?? [],
-            clipAffectsMain(src),
-            affectedSlotIds,
-            positionAt
-          )
-          const clips = animation.clips.flatMap((clip) => {
-            if (clip.id !== id) return [clip]
-            const first: AnimationClip = {
-              ...clip,
-              durationMs: atMs - clip.startMs,
-              pose: midPose,
-            }
-            const second: AnimationClip = {
-              ...clip,
-              id: newId,
-              startMs: atMs,
-              durationMs: srcEnd - atMs,
-              pose: toPose,
-            }
-            return [first, second]
-          })
-          return { animation: { ...animation, clips } }
-        },
-        null
-      )
-      // Hand the open-clip role to the second half (its pose is the live canvas,
-      // so the canvas already shows it — no reload needed). A later
-      // selectAnimationClip(newId) then no-ops instead of re-saving over the cut.
-      if (wasOpen) set({ selectedAnimationClipId: newId })
-      return newId
-    },
-    clearAnimationClips: (canvasId) =>
-      commitCanvas(
-        canvasId,
-        (canvas) => ({
-          animation: { ...getCanvasAnimation(canvas), clips: [] },
-        }),
-        null
-      ),
-    setIsPreviewMode: (p) => set({ isPreviewMode: p }),
-    setIsPreviewAutoScroll: (a) => set({ isPreviewAutoScroll: a }),
-    setPreviewAnimation: (a) => set({ previewAnimation: a }),
-    setPreviewAutoScrollDelay: (d) => set({ previewAutoScrollDelay: d }),
-    setBulkEditMode: (b) => {
-      if (!b) {
-        // Reset all canvas positions to center when disabling bulk edit
-        commit(
-          (state) => ({
-            canvases: state.canvases.map((c) => ({
-              ...c,
-              position: { x: 0, y: 0 },
-            })),
-          }),
-          null
-        )
-      }
-      set({ bulkEditMode: b, bulkCanvasDragging: false, bulkViewportZoom: 1 })
-    },
-    setBulkCanvasDragging: (dragging) => set({ bulkCanvasDragging: dragging }),
-    setScreenshotPositionDragging: (dragging) =>
-      set({ screenshotPositionDragging: dragging }),
-    setBulkViewportZoom: (zoom) =>
-      set({ bulkViewportZoom: Math.max(0.05, Math.min(2, zoom)) }),
-    setBulkScale: (n) => set({ bulkScale: Math.max(20, Math.min(100, n)) }),
-
-    undo: () => {
-      const state = get()
-      if (!state.past.length) return
-      const prev = state.past[state.past.length - 1]
-      set({
-        past: state.past.slice(0, -1),
-        present: prev,
-        future: [state.present, ...state.future],
-        _lastGroup: null,
-        _lastTs: 0,
-        bulkEditMode: prev.canvases.length > 1,
-      })
-    },
-    redo: () => {
-      const state = get()
-      if (!state.future.length) return
-      const next = state.future[0]
-      set({
-        past: [...state.past, state.present],
-        present: next,
-        future: state.future.slice(1),
-        _lastGroup: null,
-        _lastTs: 0,
-        bulkEditMode: next.canvases.length > 1,
-      })
-    },
-    reset: () => {
-      const state = get()
-      set({
-        past: [...state.past, state.present],
-        present: DEFAULT_STATE,
-        future: [],
-        _lastGroup: null,
-        _lastTs: 0,
-        bulkCanvasDragging: false,
-        bulkViewportZoom: 1,
-        currentDraft: null,
-        activeLayoutPresetId: null,
-        activeCustomPresetId: null,
-        activeSinglePresetId: null,
-      })
-    },
-    addCanvas: () => {
-      if (get().present.canvases.length >= MAX_CANVASES) return null
-      const id = makeId()
-      commit((state) => {
-        const newCanvas = createCanvas(id, placementAfterCanvas(state))
-        return {
-          canvases: [...state.canvases, newCanvas],
-          activeCanvasId: id,
-        }
-      }, null)
-      set((s) => ({
-        ...CLEAR_SELECTION,
-        bulkEditMode: true,
-        bulkCanvasDragging: false,
-        bulkViewportZoom: 1,
-        bulkFitViewSeq: s.bulkFitViewSeq + 1,
-      }))
-      return id
-    },
-    removeCanvas: (id) => {
-      commit((state) => {
-        if (state.canvases.length <= 1) return {}
-        const remaining = state.canvases.filter((c) => c.id !== id)
-        const activeCanvasId =
-          state.activeCanvasId === id
-            ? (remaining[0]?.id ?? state.activeCanvasId)
-            : state.activeCanvasId
-        return { canvases: remaining, activeCanvasId }
-      }, null)
-      set({ ...CLEAR_SELECTION })
-    },
-    duplicateCanvas: (sourceId) => {
-      if (get().present.canvases.length >= MAX_CANVASES) return null
-      const newId = makeId()
-      let didCopy = false
-      commit((state) => {
-        const targetId = sourceId ?? state.activeCanvasId
-        const srcIndex = state.canvases.findIndex((c) => c.id === targetId)
-        if (srcIndex < 0) return {}
-        didCopy = true
-        const src = state.canvases[srcIndex]
-        const copy: CanvasState = {
-          ...src,
-          id: newId,
-          position: placementAfterCanvas(state, src.id),
-        }
-        // Insert the copy right after the source canvas
-        const canvases = [...state.canvases]
-        canvases.splice(srcIndex + 1, 0, copy)
-        return {
-          canvases,
-          activeCanvasId: newId,
-        }
-      }, null)
-      if (didCopy) {
-        set((s) => ({ bulkFitViewSeq: s.bulkFitViewSeq + 1 }))
-      }
-      return didCopy ? newId : null
-    },
-    setActiveCanvasId: (id) => {
-      const state = get()
-      if (state.present.activeCanvasId === id) return
-      commit({ activeCanvasId: id }, null)
-      set({ ...CLEAR_SELECTION })
-    },
-    setCanvasPosition: (id, position) => {
-      commit(
-        (state) => ({
-          canvases: state.canvases.map((c) =>
-            c.id === id ? { ...c, position } : c
-          ),
-        }),
-        `canvasPosition-${id}`
-      )
-    },
-    setCanvasPositions: (positions) => {
-      commit(
-        (state) => ({
-          canvases: state.canvases.map((c) =>
-            positions[c.id] ? { ...c, position: positions[c.id] } : c
-          ),
-        }),
-        null
-      )
-    },
-    requestBulkFitView: () =>
-      set((s) => ({ bulkFitViewSeq: s.bulkFitViewSeq + 1 })),
-
-    addScreenshotSlot: (canvasId) => {
-      const targetId = canvasId ?? get().present.activeCanvasId
-      const target = get().present.canvases.find(
-        (canvas) => canvas.id === targetId
-      )
-      if (
-        !target ||
-        target.tweet ||
-        (target.screenshot && isVideoSrc(target.screenshot)) ||
-        target.screenshotSlots.length >= MAX_SCREENSHOT_SLOTS
-      ) {
-        return null
-      }
-      const id = makeId()
-      commitCanvas(
-        targetId,
-        (canvas, state) => {
-          const next = createScreenshotSlot(
-            {
-              id,
-              tilt: { ...canvas.tilt },
-              scale: canvas.scale,
-              border: cloneBorder(canvas.border),
-              borderRadius: canvas.borderRadius,
-              padding: canvas.padding,
-              shadow: cloneShadow(canvas.shadow),
-              lighting: cloneLighting(canvas.backdrop.lighting),
-            },
-            computeNextLayerZ(canvas)
-          )
-          return {
-            screenshotSlots: placeNewSlotInRow(
-              canvas.screenshotSlots,
-              next,
-              canvas.frame,
-              stateCanvasAspect(state)
-            ),
-          }
-        },
-        null
-      )
-      return id
-    },
-    updateScreenshotSlot: (id, patch, canvasId) => {
-      const apply = (canvas: CanvasState) => ({
-        screenshotSlots: canvas.screenshotSlots.map((slot) =>
-          slot.id === id ? { ...slot, ...patch } : slot
-        ),
-      })
-      // A slot's transform + shadow + position + border/radius/padding/lighting
-      // edits become owned effects on the open keyframe; other slot changes (fit,
-      // filter…) don't animate, so they commit normally.
-      const effects: AnimationEffect[] = []
-      if ("tilt" in patch || "rotation" in patch) effects.push("tilt")
-      if ("scale" in patch) effects.push("zoom")
-      if ("shadow" in patch) effects.push("shadow")
-      if ("xPct" in patch || "yPct" in patch) effects.push("position")
-      if ("border" in patch) effects.push("border")
-      if ("borderRadius" in patch) effects.push("borderRadius")
-      if ("padding" in patch) effects.push("padding")
-      if ("lighting" in patch) effects.push("lighting")
-      if (effects.length === 0) {
-        commitCanvas(canvasId, apply, `screenshot-slot-${id}`)
-      } else {
-        commitCanvasEffect(canvasId, apply, `screenshot-slot-${id}`, effects)
-      }
-    },
-    setScreenshotSlotImage: (id, src, canvasId) => {
-      if (src === null) {
-        commitCanvas(
-          canvasId,
-          (canvas) => ({
-            screenshotSlots: canvas.screenshotSlots.map((slot) =>
-              slot.id === id
-                ? {
-                    ...slot,
-                    src,
-                    originalSrc: null,
-                    lastCropRegion: null,
-                    fullPageCapture: null,
-                  }
-                : slot
-            ),
-          }),
-          null
-        )
-        return
-      }
-
-      const snapshot = get()
-      commitCanvas(
-        canvasId,
-        (canvas, state) => {
-          const activeLayoutGeometry = resolveActiveLayoutGeometry(
-            snapshot,
-            canvas.frame
-          )
-          const updatedSlots = canvas.screenshotSlots.map((slot) =>
-            slot.id === id
-              ? {
-                  ...slot,
-                  src,
-                  originalSrc: src,
-                  lastCropRegion: null,
-                  fullPageCapture: null,
-                  objectFit: slot.objectFit ?? "contain",
-                }
-              : slot
-          )
-          if (
-            !activeLayoutGeometry ||
-            updatedSlots.length !== activeLayoutGeometry.slots.length
-          ) {
-            return { screenshotSlots: updatedSlots }
-          }
-          // When the active layout preset uses relative slot positions, the
-          // preset's xPct/yPct are offsets from each slot's natural row-layout
-          // position — not absolute values. Mirror the same resolution
-          // applyLayoutPreset does so uploading an image doesn't snap the box
-          // to (0, 0).
-          const naturalLayout = computeRowLayout(
-            [
-              { id: "__main__", frame: canvas.frame },
-              ...updatedSlots.map((slot) => ({
-                id: slot.id,
-                frame: slot.frame ?? canvas.frame,
-              })),
-            ],
-            stateCanvasAspect(state)
-          )
-          return {
-            screenshotSlots: updatedSlots.map((slot, index) => {
-              const config = activeLayoutGeometry.slots[index]
-              if (!config) return slot
-              const { xPct, yPct } = resolveSlotPositionPct({
-                config,
-                naturalSlotXPct: naturalLayout[index + 1]?.xPct ?? slot.xPct,
-                relativeSlotPositions:
-                  activeLayoutGeometry.relativeSlotPositions,
-              })
-              return {
-                ...slot,
-                xPct,
-                yPct,
-                rotation: config.rotation,
-                tilt: config.tilt,
-                scale: config.scale,
-                ...(config.zIndex !== undefined && { zIndex: config.zIndex }),
-              }
-            }),
-          }
-        },
-        null
-      )
-    },
-    setFullPageScreenshotSlot: (id, src, canvasId) =>
-      commitCanvas(
-        canvasId,
-        (canvas) => ({
-          screenshotSlots: canvas.screenshotSlots.map((slot) =>
-            slot.id === id
-              ? {
-                  ...slot,
-                  src,
-                  originalSrc: src,
-                  lastCropRegion: null,
-                  fullPageCapture: src ? { scrollPosition: 0 } : null,
-                }
-              : slot
-          ),
-        }),
-        null
-      ),
-    setFullPageScreenshotSlotScrollPosition: (id, scrollPosition, canvasId) =>
-      commitCanvas(
-        canvasId,
-        (canvas) => ({
-          screenshotSlots: canvas.screenshotSlots.map((slot) =>
-            slot.id === id
-              ? {
-                  ...slot,
-                  fullPageCapture: slot.fullPageCapture
-                    ? {
-                        scrollPosition: Math.max(
-                          0,
-                          Math.min(100, scrollPosition)
-                        ),
-                      }
-                    : slot.fullPageCapture,
-                }
-              : slot
-          ),
-        }),
-        `full-page-slot-scroll-${id}`
-      ),
-    applyCroppedScreenshotSlot: (id, src, region, canvasId) =>
-      commitCanvas(
-        canvasId,
-        (canvas) => ({
-          screenshotSlots: canvas.screenshotSlots.map((slot) =>
-            slot.id === id
-              ? {
-                  ...slot,
-                  src,
-                  originalSrc: slot.originalSrc ?? slot.src,
-                  lastCropRegion: region,
-                  fullPageCapture: null,
-                }
-              : slot
-          ),
-        }),
-        `screenshot-slot-crop-${id}`
-      ),
-    deleteScreenshotSlot: (id, canvasId) =>
-      commitCanvas(
-        canvasId,
-        (canvas, state) => ({
-          screenshotSlots: removeSlotFromRow(
-            canvas.screenshotSlots,
-            id,
-            canvas.frame,
-            stateCanvasAspect(state)
-          ),
-        }),
-        null
-      ),
-    duplicateScreenshotSlot: (id, canvasId) => {
-      const targetId = canvasId ?? get().present.activeCanvasId
-      const target = get().present.canvases.find(
-        (canvas) => canvas.id === targetId
-      )
-      if (!target || target.screenshotSlots.length >= MAX_SCREENSHOT_SLOTS) {
-        return null
-      }
-      const copyId = makeId()
-      let didCopy = false
-      commitCanvas(
-        targetId,
-        (canvas, state) => {
-          const src = canvas.screenshotSlots.find((slot) => slot.id === id)
-          if (!src) return { screenshotSlots: canvas.screenshotSlots }
-          didCopy = true
-          const copy: ScreenshotSlot = {
-            ...src,
-            id: copyId,
-            zIndex: computeNextLayerZ(canvas),
-          }
-          return {
-            screenshotSlots: placeNewSlotInRow(
-              canvas.screenshotSlots,
-              copy,
-              canvas.frame,
-              stateCanvasAspect(state)
-            ),
-          }
-        },
-        null
-      )
-      return didCopy ? copyId : null
-    },
-    bringScreenshotSlotToFront: slotLayerOps.toFront,
-    sendScreenshotSlotToBack: slotLayerOps.toBack,
-    arrangeScreenshotSlotsInRow: (canvasId) =>
-      commitCanvas(
-        canvasId,
-        (canvas, state) => ({
-          screenshotSlots: layoutSlotsInRow(
-            canvas.screenshotSlots,
-            canvas.frame,
-            stateCanvasAspect(state)
-          ),
-        }),
-        null
-      ),
-    setScreenshotSlotGroupPosition: (position, canvasId) =>
-      // Position is animatable — register ownership on the open keyframe so
-      // group moves (and the animate Position pad's "all" path) keyframe correctly.
-      commitCanvasEffect(
-        canvasId,
-        (canvas) => {
-          if (canvas.screenshotSlots.length === 0) {
-            return { screenshotSlots: canvas.screenshotSlots }
-          }
-
-          const bounds = canvas.screenshotSlots.reduce(
-            (acc, slot) => ({
-              minX: Math.min(acc.minX, slot.xPct - slot.widthPct / 2),
-              maxX: Math.max(acc.maxX, slot.xPct + slot.widthPct / 2),
-              minY: Math.min(acc.minY, slot.yPct - slot.heightPct / 2),
-              maxY: Math.max(acc.maxY, slot.yPct + slot.heightPct / 2),
-            }),
-            {
-              minX: Number.POSITIVE_INFINITY,
-              maxX: Number.NEGATIVE_INFINITY,
-              minY: Number.POSITIVE_INFINITY,
-              maxY: Number.NEGATIVE_INFINITY,
-            }
-          )
-          const centerX = (bounds.minX + bounds.maxX) / 2
-          const centerY = (bounds.minY + bounds.maxY) / 2
-          const dx = position.xPct - centerX
-          const dy = position.yPct - centerY
-
-          return {
-            screenshotSlots: canvas.screenshotSlots.map((slot) => ({
-              ...slot,
-              xPct: clampPct(slot.xPct + dx),
-              yPct: clampPct(slot.yPct + dy),
-            })),
-          }
-        },
-        "screenshot-slot-group-position",
-        "position"
-      ),
+    ...createProjectActions(context),
+    ...createMediaActions(context),
+    ...createCanvasStyleActions(context),
+    ...createLayerActions(context),
+    ...createAnimationActions(context),
+    ...createSessionActions(context),
+    ...createCanvasActions(context),
+    ...createSlotActions(context),
   }
 })
