+++ lib/editor/store/initial-state.ts
@@ -1,0 +1,41 @@
+import { DEFAULT_STATE } from "./defaults"
+import type { EditorStoreState } from "./types"
+
+export const INITIAL_EDITOR_STORE_STATE: EditorStoreState = {
+  past: [],
+  present: DEFAULT_STATE,
+  future: [],
+  _lastGroup: null,
+  _lastTs: 0,
+  topBarPopoverOpen: false,
+  isAnimateMode: false,
+  isPreviewMode: false,
+  isPreviewAutoScroll: false,
+  previewAutoScrollDelay: 3000,
+  previewAnimation: "slide",
+  bulkEditMode: false,
+  bulkCanvasDragging: false,
+  screenshotPositionDragging: false,
+  bulkViewportZoom: 1,
+  bulkScale: 65,
+  bulkFitViewSeq: 0,
+  selectedTextId: null,
+  selectedAssetId: null,
+  selectedAnnotationShapeId: null,
+  selectedScreenshotSlotId: null,
+  isScreenshotSelected: false,
+  selectedAnimationClipId: null,
+  selectedAnimationClipIds: [],
+  presetTab: "single",
+  activeLayoutPresetId: null,
+  activeCustomPresetId: null,
+  activeSinglePresetId: null,
+  customPresets: [],
+  customPresetsLoaded: false,
+  customPresetsLoading: false,
+  customPresetsError: false,
+  customPresetsForUserId: null,
+  customPresetsSort: "latest",
+  customPresetsListSort: "latest",
+  currentDraft: null,
+}
